agent-api: 78 routes into seven routers by owner, zero behaviour change (B3)
`create_app` was 2,868 lines and held every route, so every lane that touched
agent-api touched one file — and it is where the identity boundary was hardest
to see (seam backlog B3). `api.py` is now 1,064 lines: imports, `create_app`,
the shared closures the routes are built out of, and seven `include_router`
calls.

TWO MOVES, IN THIS ORDER, AND NEITHER REWRITES A LINE OF LOGIC.

1. `api_shared.py` (1,042) — the request/response models, pure helpers and
   constants, moved verbatim. The routers cannot import `api.py`: `api.py`
   imports THEM, and a module that imports its own importer is a cycle whose
   behaviour depends on which line the interpreter reached first. Not a thing to
   put under 78 routes.

2. `routers/` — one module per owner, each a single `build(**deps) -> APIRouter`:
   health 49 · chats 362 · admin 224 · meetings 169 · scaffolds 348 ·
   friction 86 · workspaces 993.

WHY `build(**deps)` AND NOT A CONTEXT OBJECT. Every handler closed over
`create_app`'s locals — `wsr`, `settings`, `subject_of`, `_read_target` and
thirty-six more. `build()` rebinds each to THE NAME IT ALREADY HAD, so every
handler body moved byte for byte: `@app.` became `@router.` and not one
identifier inside a body changed. A `deps.wsr` rewrite would have touched every
line of all 78 handlers and made the diff unreviewable. `git diff -M` reads this
as a move, and each `build()` names what that router takes, so "what does this
router depend on" is one line to read.

THE ORDER QUESTION, ANSWERED RATHER THAN ASSUMED. Regrouping changes the order
routes register in, and FastAPI resolves first-match-wins — so this is only safe
while no two routes can match the same concrete URL under the same method. None
can, measured across all 82: `tests/test_route_table.py` asserts it on every run,
so it stays true as routes are added instead of being a property this refactor
happened to have.

PROOF. Route table before/after, captured by booting both trees and walking the
effective routes: 82 → 82, and the SORTED tables are identical on path, methods,
handler name, status_code, response_model, include_in_schema, dependencies,
response_class and the path/query/body/header parameter sets. Include prefix is
empty and no dependency was injected at include time — both checked, because
either would move every path or add an auth check that no handler's source would
show. The ordered lists differ, as intended and as the collision check makes
safe.

One trap worth recording: FastAPI 0.138 does NOT flatten `include_router` into
`app.routes` — it leaves an `_IncludedRouter` and resolves at match time. A probe
that reads `app.routes` finds seven placeholders and no routes, and would have
reported this refactor green while comparing nothing. Both the probe and the test
walk through it.

Agent suite: 1615 passed, 0 failed (1612 + the 3 new route-table tests). NO TEST
WAS EDITED — not even an import path. 14 gates green.

PRD 40.7 note, in `routers/README.md`: `friction.py` is its own router because
where the friction ledger belongs is the founder's open question right now — a
`no-agents` deployment still has people filing rough edges. `scaffolds.py` is the
other candidate: minted by flows, read by the terminal, composing an agent turn.
The split is drawn so that moving either is a file move, not a grep. This commit
decides neither.
