+++ core/agent/control_plane/api.py
@@ -29,7 +29,7 @@ import time
 from pathlib import Path
 from typing import Callable, Iterator, Optional
 
-from fastapi import Body, FastAPI, File, HTTPException, Request, UploadFile
+from fastapi import APIRouter, Body, FastAPI, File, HTTPException, Request, UploadFile
 from fastapi.responses import JSONResponse, Response, StreamingResponse
 from jsonschema.exceptions import ValidationError
 from pydantic import BaseModel
@@ -91,957 +91,29 @@ from control_plane.events import event_to_invocation
 from shared.ports import SchedulerPort, StreamReader
 from control_plane.workspace_reader import WorkspaceReader
 
-logger = logging.getLogger("agent_api.api")
-
-# The phase, in the meeting's own vocabulary — the same three words the terminal's header uses
-# (`minutes/MinutesShell.tsx` PHASE_WORD). One vocabulary, two renderers.
-_PHASE_WORD = {"prep": "upcoming", "live": "live", "post": "held"}
-
-
-def _iso(epoch) -> "str | None":
-    """An epoch as an ISO-8601 UTC string, or None. The wire carries times as STRINGS because the
-    client half pins them that way; the store keeps the float, which is what arithmetic wants."""
-    try:
-        return time.strftime("%Y-%m-%dT%H:%M:%SZ", time.gmtime(float(epoch)))
-    except (TypeError, ValueError):
-        return None
-
-
-def _provenance_line(prov: dict, minted_at) -> str:
-    """The record's provenance as ONE readable line — what produced this touch, and when.
-
-    The OBJECT stays on the wire beside it: flow, reaction id, run id and minted_by are four facts,
-    and a string that concatenates four facts is not a record of them. This line exists so a panel
-    can show something without parsing, not so the object can be dropped."""
-    if not isinstance(prov, dict):
-        return ""
-    bits = [str(prov[k]) for k in ("flow", "reaction_id", "run_id") if prov.get(k)]
-    if prov.get("minted_by"):
-        bits.append(f"minted by {prov['minted_by']}")
-    stamp = _iso(minted_at)
-    if stamp:
-        bits.append(stamp)
-    return " · ".join(bits)
-
-
-def _epoch_text(when) -> str:
-    """An epoch as a readable UTC line, or "" — the last-resort rendering of a meeting's time when
-    the caller did not render one in the recipient's own zone. UTC and SAID to be UTC: a bare
-    "14:00" in a zone nobody named is the kind of half-fact that reads as a bug in an inbox."""
-    try:
-        return time.strftime("%a %d %b %H:%M UTC", time.gmtime(float(when)))
-    except (TypeError, ValueError):
-        return ""
-
-MAX_UPLOAD_BYTES = 25 * 1024 * 1024
-MEETING_STREAM_TRANSCRIPT_REPLAY = 80
-
-
-def _upload_filename(name: str | None) -> str:
-    base = (name or "upload").replace("\\", "/").rsplit("/", 1)[-1].strip()
-    base = re.sub(r"\s+", "_", base)
-    base = re.sub(r"[^A-Za-z0-9._-]", "_", base).strip("._-")
-    return base[:160] or "upload"
-
-
-def _truncate_title(text: str, *, limit: int = 60) -> str:
-    """A session's default title — the first prompt, single-lined + truncated."""
-    title = " ".join((text or "").split())
-    return title[: limit - 1] + "…" if len(title) > limit else title
-
-
-def _stream_tail_id(redis_url: str | None, stream: str) -> str | None:
-    if not redis_url:
-        return None
-    try:
-        import redis
-
-        r = redis.from_url(redis_url, decode_responses=True)
-        rows = r.xrevrange(stream, count=1)
-        return str(rows[0][0]) if rows else "0-0"
-    except Exception as exc:
-        logger.warning("could not resolve transcript stream tail for %s: %s", stream, exc)
-        return None
-
-
-# How long a turn's start-cursor record lives — covers the client's whole resume window (its hard
-# timeout is minutes); after this a stale nonce simply falls back to the fresh-dispatch path.
-CHAT_TURN_HEAD_TTL_SEC = 900
-
-
-def _chat_turn_head_key(unit_id: str) -> str:
-    return f"unit:{unit_id}:turnhead"
-
-
-def _record_chat_turn_head(redis_url: str | None, unit_id: str, turn_id: str, start: str) -> None:
-    """Remember, per warm chat unit, the CURRENT turn's nonce + the out-Stream id it started after.
-    Best-effort (redis-less unit tests skip it): losing the record only degrades a no-cursor retry
-    back to today's behavior, it never breaks the turn."""
-    if not redis_url:
-        return
-    try:
-        import redis
-
-        r = redis.from_url(redis_url, decode_responses=True)
-        r.set(_chat_turn_head_key(unit_id), json.dumps({"turn_id": turn_id, "start": start}),
-              ex=CHAT_TURN_HEAD_TTL_SEC)
-    except Exception as exc:  # noqa: BLE001
-        logger.warning("could not record chat turn head for %s: %s", unit_id, exc)
-
-
-def _chat_turn_head(redis_url: str | None, unit_id: str, turn_id: str) -> str | None:
-    """The recorded start cursor of the turn ``turn_id`` — or None when it isn't the current turn."""
-    if not redis_url or not turn_id:
-        return None
-    try:
-        import redis
-
-        r = redis.from_url(redis_url, decode_responses=True)
-        raw = r.get(_chat_turn_head_key(unit_id))
-        head = json.loads(raw) if raw else None
-        return head["start"] if head and head.get("turn_id") == turn_id else None
-    except Exception as exc:  # noqa: BLE001
-        logger.warning("could not read chat turn head for %s: %s", unit_id, exc)
-        return None
-
-
-class _Sessions:
-    """Durable, per-subject chat-session index. Each session carries a created + last-active stamp and an
-    optional title (default the first prompt, truncated). ``list`` returns them most-recent first.
-
-    Backed by redis when a client is wired (one hash per session under ``agent:sessions:<subject>`` +
-    the per-subject id set), with an in-memory fallback so the unit tests need no redis. Multiple
-    conversation threads live in the ONE user workspace — this indexes the threads, not workspaces."""
-
-    def __init__(self, redis_client=None) -> None:
-        self._redis = redis_client
-        self._mem: dict[str, dict[str, dict]] = {}  # subject → {session → {created,last_active,title}}
-
-    # ── redis key helpers ──
-    @staticmethod
-    def _ids_key(subject: str) -> str:
-        return f"agent:sessions:{subject}"
-
-    @staticmethod
-    def _meta_key(subject: str, session: str) -> str:
-        return f"agent:session:{subject}:{session}"
-
-    def _now(self) -> float:
-        import time
-
-        return time.time()
-
-    def upsert(self, subject: str, session: str, *, title: str | None = None) -> None:
-        """Record the session on use: create it (stamping ``created`` + a default ``title``) or touch its
-        ``last_active``. An explicit ``title`` overrides; otherwise the first prompt seeds it once."""
-        now = self._now()
-        if self._redis is not None:
-            mkey = self._meta_key(subject, session)
-            existing = self._redis.hgetall(mkey) or {}
-            fields = {"last_active": str(now)}
-            if not existing:
-                fields["created"] = str(now)
-                fields["title"] = title or session
-            elif title is not None:
-                fields["title"] = title
-            self._redis.hset(mkey, mapping=fields)
-            self._redis.sadd(self._ids_key(subject), session)
-            return
-        rec = self._mem.setdefault(subject, {}).get(session)
-        if rec is None:
-            self._mem[subject][session] = {"created": now, "last_active": now, "title": title or session}
-        else:
-            rec["last_active"] = now
-            if title is not None:
-                rec["title"] = title
-
-    def list(self, subject: str) -> list[dict]:
-        """The subject's sessions, most-recently-active first."""
-        rows: list[dict] = []
-        if self._redis is not None:
-            for session in self._redis.smembers(self._ids_key(subject)) or set():
-                meta = self._redis.hgetall(self._meta_key(subject, session)) or {}
-                rows.append({
-                    "session": session,
-                    "title": meta.get("title") or session,
-                    "created": float(meta.get("created", 0) or 0),
-                    "last_active": float(meta.get("last_active", 0) or 0),
-                })
-        else:
-            for session, meta in self._mem.get(subject, {}).items():
-                rows.append({
-                    "session": session, "title": meta.get("title") or session,
-                    "created": meta.get("created", 0.0), "last_active": meta.get("last_active", 0.0),
-                })
-        rows.sort(key=lambda r: r["last_active"], reverse=True)
-        return rows
-
-    def drop(self, subject: str, session: str) -> None:
-        if self._redis is not None:
-            self._redis.srem(self._ids_key(subject), session)
-            self._redis.delete(self._meta_key(subject, session))
-            return
-        self._mem.get(subject, {}).pop(session, None)
-
-
-# P21 (ADR 0027 family — the panel's stale-live finding): a registry entry is "live" only while
-# segments actually FLOW. The watcher re-adds on every batch (~2s apart), so this much silence
-# means the meeting is over even when the session_end frame was lost (e.g. a hot-reload racing the
-# wire leaves it pending-unacked) — the server-side stale-"live" the terminal's durableTerminal
-# guard was papering over.
-LIVE_SILENCE_TTL_SEC = 60.0
-
-class _LiveMeetings:
-    """In-memory registry of live meetings — the terminal's 'meetings' feed. Keyed by session_uid (the
-    native Meet code). A stopped/ended meeting is KEPT (``status='stopped'``) so the terminal can offer to
-    send the bot back; ``add`` (re)marks it live. Liveness is EVIDENCE, not a latch: ``add`` stamps
-    ``last_seen`` and ``list`` demotes an entry silent past LIVE_SILENCE_TTL_SEC to stopped (P21 —
-    absence of the expected signal is itself a reportable state). The dev-tier foundation."""
-
-    def __init__(self) -> None:
-        self._by_uid: dict[str, dict] = {}
-
-    def add(self, meeting: dict) -> None:
-        m = dict(meeting)
-        m["status"] = "live"
-        m["last_seen"] = time.monotonic()
-        self._by_uid[meeting["session_uid"]] = m
-
-    def stop(self, session_uid: str) -> None:
-        m = self._by_uid.get(session_uid)
-        if m:
-            m["status"] = "stopped"
-
-    def drop(self, session_uid: str) -> None:
-        # the meeting ended — keep the row (stopped) so 'send the bot back' stays available
-        self.stop(session_uid)
-
-    def list(self) -> list[dict]:
-        now = time.monotonic()
-        for m in self._by_uid.values():
-            if m.get("status") == "live" and now - m.get("last_seen", now) > LIVE_SILENCE_TTL_SEC:
-                m["status"] = "stopped"  # earned liveness expired — the segment flow went silent
-        return list(self._by_uid.values())
-
-
-class ChatContextBody(BaseModel):
-    """The terminal-state CONTEXT BUNDLE (slice 1). ``extra="ignore"`` on purpose — forward-
-    tolerant: a newer terminal adding bundle fields must never 422 against this server."""
-    model_config = {"extra": "ignore"}
-    tz: Optional[str] = None            # IANA tz for digest rendering (invalid → UTC)
-    surface: Optional[dict] = None      # {list?: str, tab?: {kind: str}} — the ambient gate signal
-    focus: Optional[dict] = None        # the focused thing (meeting/file/workspace/today); None = cleared
-    include: Optional[dict] = None      # {schedule?: bool} — explicit user toggle beats the gate
-
-
-class ChatBody(BaseModel):
-    model_config = {"extra": "forbid"}
-    prompt: str
-    # subject is DERIVED server-side from X-User-Id (P20) — kept here only so a client that still sends it
-    # doesn't 422 (extra=forbid); the value is IGNORED. Dropped from the client in Stage 4.
-    subject: Optional[str] = None
-    session: Optional[str] = None
-    # LEGACY single-focus grounding ({kind, ref}) — still honored when ``context`` is absent, so
-    # old clients keep byte-identical behavior. The terminal now sends ``context`` (below) too.
-    active: Optional[dict] = None
-    # the terminal-state context bundle; when present it is AUTHORITATIVE (including
-    # ``focus: null`` = the user cleared the focus chip — legacy ``active`` is then ignored).
-    context: Optional[ChatContextBody] = None
-    # Client-minted TURN NONCE (one per user turn, constant across that turn's reconnect attempts).
-    # Lets the server tell a no-cursor RETRY (the stream dropped before the client ever saw an ``id:``,
-    # so it can't send Last-Event-ID) from a genuinely new turn with identical text ("yes" twice):
-    # a matching nonce re-attaches from the turn's recorded start — no second dispatch, no lost events.
-    turn_id: Optional[str] = None
-    # THE MEETING ROOM (post-meeting run). The caller may name ONLY THE MEETING — a meetings-domain
-    # ROW id — and the server resolves who was in it (control_plane/meeting_room.py). There is
-    # deliberately NO field here that names a workspace or a subject: a caller able to say "mount
-    # u_bob" could read any user's desk by naming it, which is the exact hole this shape exists to
-    # close. Accepting it additionally requires the internal-tier secret (see ``_resolve_room``), so
-    # the general end-user chat surface cannot open a room at all.
-    room_meeting_id: Optional[str] = None
-    # MEMBERSHIP = the INVITE's participant list, as ADDRESSES. The trusted caller holds the parsed
-    # invite, so it sends the addresses; agent-api resolves each one to a subject through admin-api
-    # and mounts only participants who ALREADY have a subject and a desk. Addresses, never subject
-    # ids and never workspace names: the resolution — and therefore the blast radius — stays here.
-    room_participants: Optional[list[str]] = None
-    # The invite's ICS ``CN=`` map (address → display name). Used ONLY to match transcript speaker
-    # labels back to addresses that are ALREADY in the list above, i.e. only to ORDER the room. A
-    # name never admits anybody, so a bad match costs position and nothing else.
-    room_participant_names: Optional[dict[str, str]] = None
-    # The transcript's speaker labels, already ordered by speaking time DESCENDING (the caller holds
-    # the transcript, so it does that arithmetic). Unmatched labels are simply ignored.
-    room_speakers: Optional[list[str]] = None
-    room_read_max: Optional[int] = None
-    # THE SCAFFOLD (PRD 5.5). The terminal sends this on the FIRST turn of a chat a link composed.
-    # It is an ID, never a mount list and never prompt text: the record it names says which
-    # workspaces this turn mounts and what the opening ask is, and the server reads BOTH from the
-    # store. A scaffold that is not this subject's is ignored (never an error) — a stale or
-    # forwarded id must not be able to widen anybody's mounts, and must not break their turn either.
-    scaffold_id: Optional[str] = None
-    # THE INTENT (PRD decision 32/35). A button pressed on a page — Extend, Create this page,
-    # Explore a term in a transcript, Highlight the transcript — is an ACT on a named thing, not a
-    # sentence somebody typed. The terminal has sent this since decision 32 landed client-side; with
-    # `extra="forbid"` above and no field here, every one of those presses 422'd. It is a plain dict
-    # rather than a model on purpose: the CLOSED vocabulary is `chat_intents.INTENT_PRESETS`, an
-    # unknown kind is ignored, and a client one release ahead must not be refused at the door.
-    intent: Optional[dict] = None
-
-
-class ScaffoldMintBody(BaseModel):
-    """`POST /internal/scaffolds` — what a FLOW says at the moment it creates a touch.
-
-    ``extra="forbid"``: a mint is the thing a step checks before it sends, so a caller that spells a
-    field wrong has to hear about it here rather than mail a link built from a field nobody read.
-
-    There is deliberately NO field carrying prompt text. ``opening`` is a NAME in `_global/asks/`
-    and the server refuses anything that is not one — the URL never carried text (PRD 6) and neither
-    does the record behind it, for the same reason: anyone who can mint a touch would otherwise be
-    able to drive the recipient's agent."""
-    model_config = {"extra": "forbid"}
-    who: str                                   # the RECIPIENT ADDRESS. Not a subject: they may not exist yet.
-    kind: str                                  # one of scaffolds.KINDS
-    opening: str                               # a preset NAME in _global/asks/, never text
-    meeting: Optional[str] = None              # the meetings-domain ROW id, or None. PHASE IS NOT STORED.
-    workspaces: Optional[list[str]] = None     # slugs to mount; None = derived (see the route)
-    refs: Optional[dict] = None                # the facts for the agent: title, when, organizer, participants…
-    tabs: Optional[list[str]] = None           # UI half; None = the preset's own `tabs:`
-    focus: Optional[str] = None                # UI half; None = the preset's own `focus:`
-    # The RESTRICTED transcript share, minted by the caller against the meeting ROW
-    # (`flows_steps/meeting.mint_transcript_share`) when the meeting is not the recipient's own.
-    # Minted THERE and not here because the mint needs the meeting OWNER's gateway key, which flows
-    # holds and agent-api deliberately does not — see the route's docstring.
-    share_token: Optional[str] = None
-    provenance: Optional[dict] = None          # flow, reaction/run id, minted_by (the fact's admitted_by)
-
-
-class ScaffoldHandBody(BaseModel):
-    """`POST /api/scaffolds/hand` — a HAND LINK (`/?ask=<preset>&meeting=<row>`) turned into a record.
-
-    Two fields, both NAMES, neither prompt text — the same rule as the internal mint and for the same
-    reason (PRD 6, decisions 13/18): a URL must never be able to drive somebody's agent. What made
-    this route necessary is that the terminal used to substitute `?meeting=`/`?ws=` straight into the
-    composed opening, so a crafted link put attacker-chosen text into the first turn.
-
-    There is deliberately NO `who`: the recipient is the SIGNED-IN CALLER, taken from the session.
-    A field would let anyone who can reach this route mint a first turn for somebody else."""
-    model_config = {"extra": "forbid"}
-    preset: str                                # a preset NAME in _global/asks/, never text
-    meeting: Optional[str] = None              # a meetings-domain ROW id the CALLER can see, or None
-
-
-class ResetBody(BaseModel):
-    """Body for POST /api/chat/reset — the docs (api/agent.mdx) say it's just ``{session?}``. reset only
-    needs the session; ``prompt``/``subject``/``active`` are accepted-and-ignored so a client reusing the
-    chat-body shape doesn't 422 (reset must NOT require a prompt the way the chat turn does)."""
-    model_config = {"extra": "forbid"}
-    session: Optional[str] = None
-    subject: Optional[str] = None
-    prompt: Optional[str] = None
-    active: Optional[dict] = None
-    context: Optional[ChatContextBody] = None  # accepted-and-ignored, same rationale
-    room_meeting_id: Optional[str] = None      # accepted-and-ignored (reset mounts nothing)
-
-
-class RoutineCreate(BaseModel):
-    """The Routines surface / ``/routine`` create form — compiles to a routine.v1 + a schedule.v1 job."""
-    model_config = {"extra": "forbid"}
-    subject: Optional[str] = None  # DERIVED from X-User-Id (P20); ignored if sent. Dropped client-side in Stage 4.
-    name: str
-    cron: str
-    prompt: str
-    run_now: bool = True  # fire one immediate run so the author sees a result without waiting for cron
-
-
-class RoutineEnabledPatch(BaseModel):
-    model_config = {"extra": "forbid"}
-    enabled: bool
-
-
-class WorkspaceSwapBody(BaseModel):
-    """Attach a custom external git repo as the subject's workspace. Omit ``repo`` to swap back to seed."""
-    model_config = {"extra": "forbid"}
-    repo: Optional[str] = None   # git URL to clone (None → swap back to the seeded default)
-    ref: Optional[str] = None    # branch/tag/sha to check out (defaults to main)
-    slug: Optional[str] = None   # target a parked slot DIRECTLY (e.g. a no-repo backup) — restores, no re-clone
-    fresh: bool = False          # swap-to-seed only: rebuild the default from template (start fresh) vs restore the park
-    token: Optional[str] = None  # access token for a PRIVATE repo — used for the clone only, never stored (P15)
-
-
-class WorkspacePublishBody(BaseModel):
-    """Publish the subject's vexa-born workspace to GitHub — create the repo (unless ``remote_url``
-    targets a pre-created one) and push the current branch's full history. ``token`` is the caller's
-    PAT, used server-side for this call only, NEVER stored (P15)."""
-    model_config = {"extra": "forbid"}
-    repo_name: Optional[str] = None    # name of the repo to create (required unless remote_url is given)
-    private: bool = True               # create the repo private (default) or public
-    token: Optional[str] = None        # GitHub PAT (repo-creation + push); OPTIONAL — falls back to the caller's SAVED token
-    org: Optional[str] = None          # create under this org instead of the user's account
-    remote_url: Optional[str] = None   # skip creation and push to this (pre-created/empty) repo
-    slug: Optional[str] = None         # target workspace (own slot or shared membership); omitted = the seed-slot workspace
-
-
-class WorkspaceRenameBody(BaseModel):
-    """Set a workspace slot's DISPLAY name (label only — the slug/parked dir are unchanged). Empty clears it."""
-    model_config = {"extra": "forbid"}
-    slug: str
-    name: Optional[str] = None
-
-
-class WorkspacePushBody(BaseModel):
-    """Push a workspace's current branch to its GitHub home (origin / vexa-publish), fast-forward only.
-    ``slug`` targets one of the caller's workspaces (default = the primary); ``token`` is the caller's PAT.
-    OPTIONAL — when omitted, the caller's SAVED reusable GitHub token (git_credentials) is used. Whichever
-    token applies is used for this push only and NEVER stored on the workspace remote (P15)."""
-    model_config = {"extra": "forbid"}
-    slug: Optional[str] = None
-    token: Optional[str] = None
-
-
-class GitTokenBody(BaseModel):
-    """Save (or, with an empty/omitted ``token``, CLEAR) the caller's reusable GitHub token — stored ONCE,
-    server-side, and reused as the fallback credential for every git op across all their repos."""
-    model_config = {"extra": "forbid"}
-    token: Optional[str] = None
-
-
-class WorkspacePullBody(BaseModel):
-    """Fetch + fast-forward a workspace from its GitHub home. ``slug`` targets one of the caller's
-    workspaces (default = primary); ``token`` (optional — public repos need none) is used for the fetch
-    only and NEVER stored (P15). A divergence is refused, not merged/rebased/forced."""
-    model_config = {"extra": "forbid"}
-    slug: Optional[str] = None
-    token: Optional[str] = None
-
-
-class WorkspacePurposeBody(BaseModel):
-    """Set a workspace's PURPOSE — a one-line statement of what it's for, stored IN the workspace so it
-    travels when shared and is read into the agent's mount preamble. ``slug`` targets one of the caller's
-    workspaces (default = primary); an empty ``purpose`` clears it."""
-    model_config = {"extra": "forbid"}
-    slug: Optional[str] = None
-    purpose: str = ""
-
-
-class InviteCreateBody(BaseModel):
-    """Mint a scoped invite for a shared workspace (owner/contributor only). Returns the token ONCE."""
-    model_config = {"extra": "forbid"}
-    workspace_id: str
-    role: str = "viewer"                 # viewer | contributor (never owner)
-    expires_in_sec: int = 604800         # 7 days
-    max_uses: int = 1
-    mode: str = "open"                   # open (anyone-with-link) | restricted (allowed_emails only)
-    allowed_emails: Optional[list[str]] = None  # restricted mode: the verified emails permitted to redeem
-
-
-class InviteAcceptBody(BaseModel):
-    """Redeem an invite token (any logged-in user). Idempotent per user."""
-    model_config = {"extra": "forbid"}
-    token: str
-
-
-class RoleSetBody(BaseModel):
-    """Flip a member's role (owner only) — the "change read/write permissions" DoD item."""
-    model_config = {"extra": "forbid"}
-    role: str                            # viewer | contributor | owner
-
-
-class SharedNewBody(BaseModel):
-    """CREATE a new shared workspace (top-level, caller becomes owner) — the bootstrap that makes a
-    workspace shareable so invites can be minted against it. ``name`` → display + workspace-id base."""
-    model_config = {"extra": "forbid"}
-    name: str = "Shared workspace"
-
-
-class SharedAttachBody(BaseModel):
-    """LOAD AN EXISTING REPO into a SHARED (group) workspace — the group counterpart of
-    ``POST /api/workspace/swap``. The group's current tree is PARKED (kept, swappable-back), the repo is
-    cloned in, and the member list is carried across so nobody loses access.
-
-    ``token`` is the terminal's optional per-call PAT for an https repo; the MCP path never sends one —
-    an ssh repo authenticates with the workspace's deploy key, resolved server-side."""
-    model_config = {"extra": "forbid"}
-    repo: Optional[str] = None   # git URL (ssh → deploy key; https → PAT). None + ``slug`` = swap back
-    ref: Optional[str] = None    # branch/tag/sha (defaults to main)
-    slug: Optional[str] = None   # a parked slot to restore DIRECTLY (no re-clone), e.g. "seed"
-    token: Optional[str] = None  # https only, used for the clone and never stored (P15)
-
-
-class SharedActiveBody(BaseModel):
-    """Switch a shared workspace ON (mount) or OFF (hide) in the caller's active set — per-user, membership
-    is unchanged."""
-    model_config = {"extra": "forbid"}
-    active: bool
-
-
-class ArchiveBody(BaseModel):
-    """Archive (collapse, keep) or un-archive one of the caller's own workspaces."""
-    model_config = {"extra": "forbid"}
-    archived: bool = True
-
-
-class WorkspaceActivateBody(BaseModel):
-    """ADD a workspace to the subject's active set (the additive mount set — WP-A2.1). Pass ``repo`` to
-    clone/restore a git repo, or ``slug`` to activate an already-parked slot. Unlike swap it does NOT park
-    the others — the private baseline and any other active workspaces stay mounted."""
-    model_config = {"extra": "forbid"}
-    repo: Optional[str] = None   # git URL to clone (first time) / restore (thereafter)
-    ref: Optional[str] = None    # branch/tag/sha (defaults to main)
-    slug: Optional[str] = None   # activate an already-parked slot directly (no repo needed)
-    token: Optional[str] = None  # access token for a PRIVATE repo — clone only, never stored (P15)
-
-
-class WorkspaceNewBody(BaseModel):
-    """CREATE a brand-new BLANK workspace (seeded from the template) at a fresh slug and ADD it to the
-    active set — the additive-model "new workspace" action. NOT a swap: nothing is parked/rebuilt/backed
-    up. ``name`` (optional) → the new workspace's display label (default a unique "New workspace")."""
-    model_config = {"extra": "forbid"}
-    name: Optional[str] = None
-
-
-class WorkspaceDeactivateBody(BaseModel):
-    """REMOVE a workspace from the active set (park it — never destroyed). The private baseline cannot be
-    deactivated (it is the subject's durable memory root)."""
-    model_config = {"extra": "forbid"}
-    slug: str
-
-
-def _encode_sse_cursor(last: dict, tkey: str) -> str:
-    """The transcript stream's redis cursor as the SSE event id (the browser echoes it as
-    Last-Event-ID on reconnect → we resume EXACTLY from here, gapless). '-' = not-yet-read.
-    ONE part since PRD decision 34: the copilot out-stream and the processed-notes stream that
-    used to occupy the other two are gone, and with them everything they carried."""
-    return str(last.get(tkey, "-"))
-
-
-def _decode_sse_cursor(raw: str | None) -> "str | None":
-    """Last-Event-ID → the transcript stream id, or None when absent/malformed (fresh connect).
-    Tolerates the retired multi-part form (``transcript|output|processed``) a client reconnecting
-    across the deploy still holds: the first field was always the transcript cursor."""
-    if not raw:
-        return None
-    head = raw.split("|")[0].strip()
-    return head if head and head != "-" else None
-
-
-def _sse(events) -> Iterator[str]:
-    for item in events:
-        # ``None`` is the reader's idle tick → an SSE comment keepalive. Proxies (and the client's own
-        # idle-stall detector) cut a byte-silent stream in ~18-30s; a long agent think is byte-silent for
-        # minutes. The comment is invisible to EventSource parsing but keeps bytes flowing.
-        if item is None:
-            yield ": keepalive\n\n"
-            continue
-        # Each item is either a bare event dict, or (event, sse_id) — the id makes reconnects resumable.
-        ev, sid = item if isinstance(item, tuple) else (item, None)
-        prefix = f"id: {sid}\n" if sid else ""
-        yield f"{prefix}data: {json.dumps(ev)}\n\n"
-
-
-def _has_custom_model_endpoint(cfg: dict) -> bool:
-    """True iff a per-user Settings → Models config actually delivers a credential to the worker.
-
-    ONE spelling, shared with the dispatch overlay and the Test button (F93): this used to be a
-    third hand-written copy of "is this custom", and the three disagreed — the Test button certified
-    a configuration the turn would not use. It now delegates to `model_endpoint.has_custom_endpoint`
-    so the pre-flight gate and the dispatch cannot answer differently."""
-    return model_endpoint.has_custom_endpoint(cfg)
-
-
-def _model_creds_error_message() -> str:
-    keys = ", ".join(missing_capability_keys("model_inference"))
-    return (
-        "No model credentials are configured, so the agent cannot run. "
-        f"Set one of {keys} in the deployment environment "
-        "(deploy/compose/.env for the compose stack, then `make all`), "
-        "or add a custom endpoint under Settings → Models."
-    )
-
-
-MEETING_CHAT_TRANSCRIPT_SEGMENTS = 400  # bound the live transcript folded into a meeting-chat prompt
-
-
-def _fold_meeting_transcript(redis_url: "str | None", stream_key: str, *, limit: int) -> str:
-    """Fold the live transcript Stream ``tc:meeting:{stream_key}`` — the SAME stream the terminal
-    renders — into ordered ``speaker: text`` lines for chat
-    grounding. ``stream_key`` is the meetings-domain ROW id (P0 cross-tenant leak fix: the carrier keys
-    on the row id, never the native id which collides across tenants/re-sends). Refining live drafts are
-    upserted by ``segment_id`` (latest text wins, no duplicate), arrival order preserved, bounded to the
-    last ``limit`` segments. Best-effort: returns "" when redis is unwired or the stream is empty."""
-    if not redis_url:
-        return ""
-    try:
-        import redis
-
-        r = redis.from_url(redis_url, decode_responses=True)
-        rows = r.xrange(f"tc:meeting:{stream_key}")
-    except Exception as exc:  # noqa: BLE001 — grounding is best-effort; never fail the chat turn
-        logger.warning("could not read transcript for %s: %s", stream_key, exc)
-        return ""
-    order: list[str] = []
-    seg_by_id: dict[str, dict] = {}
-    for entry_id, fields in rows:
-        payload = json.loads(fields.get("payload", "{}"))
-        if payload.get("type") == "session_end":
-            continue
-        for i, seg in enumerate(payload.get("segments", [])):
-            sid = str(seg.get("segment_id") or f"{entry_id}:{i}")
-            if sid not in seg_by_id:
-                order.append(sid)
-            seg_by_id[sid] = seg
-    lines: list[str] = []
-    for sid in order[-limit:]:
-        seg = seg_by_id[sid]
-        text = (seg.get("text") or "").strip()
-        if not text:
-            continue
-        speaker = (seg.get("speaker") or "Speaker").strip()
-        lines.append(f"{speaker}: {text}")
-    return "\n".join(lines)
-
-
-def _meeting_grounding(
-    active: "dict | None", session: str, prompt: str, redis_url: "str | None"
-) -> "tuple[dict, list[str], str]":
-    """Cookbook #1 — chat grounding in the terminal's ACTIVE meeting, branched by the meeting's
-    LIFECYCLE PHASE (design-spec meeting-lifecycle-v2, W4; steering templates + the _global override
-    live in control_plane/meeting_steering.py):
-
-      prep (idle/scheduled)          — no transcript fold (none exists); steer toward preparation,
-                                       naming the bound prep workspace when the client sent one.
-      live (default; absent status)  — fold ``tc:meeting:{row}`` fresh on every turn — a legacy
-                                       client that sends no status keeps exactly this behavior.
-      post (completed/failed/stopped)— fold the recorded transcript ``tc:meeting:{row}``; if there is
-                                       none say so plainly (fail loud, never fabricate).
-
-    There is ONE transcript and one fold of it: the "processed notes" the post branch used to prefer
-    were the in-product inference pipeline's output, and PRD decision 34 removed that pipeline.
-    Returns the plain (none-context, no tools, prompt) when the active tab isn't a meeting."""
-    a = active or {}
-    if a.get("kind") != "meeting":
-        return ({"kind": "none", "session": session}, [], prompt)
-    m = a.get("meeting") or a  # tolerate {kind, meeting:{…}} or a flat {kind, platform, native_id}
-    native = m.get("native_id") or m.get("ref")
-    if not native:
-        return ({"kind": "none", "session": session}, [], prompt)
-    platform = m.get("platform") or "google_meet"
-    # A chat turn (trigger "message"), not a live-meeting serve — the transcript travels in the prompt,
-    # so the dispatch context stays plain (no meeting env / serve path is engaged for a chat).
-    ctx = {"kind": "none", "session": session}
-    # P0 (cross-tenant leak fix): read the streams by the meetings-domain ROW id (``meeting_id``),
-    # which the terminal passes on the active meeting — the carriers key on it, never the native id
-    # (which would fold a DIFFERENT tenant's / an older row's transcript into this user's chat).
-    # Fall back to native only when the client didn't send a row id (legacy), documented as best-effort.
-    stream_key = str(m.get("meeting_id") or native)
-    status = str(m.get("status") or "").strip().lower()
-    phase = meeting_steering.phase_for(status)
-    fields = {
-        "title": str(m.get("title") or "").strip() or str(native),
-        "platform": platform,
-        "native": str(native),
-        "meeting_id": str(m.get("meeting_id") or native),   # the ROW id — the meeting's stable ?meeting= link
-    }
-
-    if phase == "prep":
-        when = str(m.get("scheduled_at") or "").strip()
-        fields["when"] = f", scheduled for {when}" if when else " (no time set yet)"
-        workspace = str(m.get("workspace_id") or "").strip()
-        fields["workspace"] = (
-            f"The prep workspace bound to this meeting is \"{workspace}\" — ground your research and "
-            f"write the brief there (its kg/ entities cover the attendees and companies). "
-            if workspace
-            else (
-                "No shared prep workspace is bound to this meeting — the brief lives (or will live) "
-                "as this meeting's note under kg/entities/meeting/ in the user's OWN workspace, "
-                "reused across the meeting's series. "
-            )
-        )
-        return (ctx, [], meeting_steering.render("prep", fields) + prompt)
-
-    if phase == "post":
-        fields["failed"] = " — the bot FAILED during this meeting" if status == "failed" else ""
-        folded = _fold_meeting_transcript(redis_url, stream_key, limit=MEETING_CHAT_TRANSCRIPT_SEGMENTS)
-        fields["source"] = "raw transcript"
-        if not folded:
-            return (ctx, [], meeting_steering.NO_RECORD_POST.format(**fields) + prompt)
-        fields["transcript"] = folded
-        return (ctx, [], meeting_steering.render("post", fields) + prompt)
-
-    transcript = _fold_meeting_transcript(redis_url, stream_key, limit=MEETING_CHAT_TRANSCRIPT_SEGMENTS)
-    if transcript:
-        fields["transcript"] = transcript
-        preamble = meeting_steering.render("live", fields)
-    else:
-        preamble = meeting_steering.NO_TRANSCRIPT_LIVE.format(platform=platform, native=native)
-    return (ctx, [], preamble + prompt)
-
-
-# The grounding/user-message boundary marker. Every server-folded context block (kg-links + mounts,
-# added by the WORKER, land BEFORE the user prompt too; schedule digest; meeting/workspace grounding)
-# sits BEFORE this sentinel; the user's actual words are AFTER it. The terminal strips everything up to
-# and including it in ONE cut — robust to preamble wording drift, unlike per-block regexes. An HTML
-# comment so the model treats it as inert. Old chats (no sentinel) fall back to the client's regexes.
-CONTEXT_SENTINEL = "<!--vexa:user-input-below-->"
-
-
-# ── terminal-state context bundle (slice 1) — the grounding orchestrator ────────────────────────────
-# A chat turn's prompt is assembled [ambient <schedule> digest] + [focus fold] + user prompt.
-#   ambient — the schedule digest, SURFACE-GATED (Meetings list / Today tab / meeting-ish tab focused)
-#             with the user's explicit include.schedule toggle beating the gate either way.
-#   focus   — meeting/prep (delegates to _meeting_grounding, ENRICHED with the server row so a cold
-#             client store can't ground a planned meeting as live), workspace (purpose + README head),
-#             today (the full-day digest REPLACES ambient), file/none (unchanged).
-# Everything stays inside the trusted control plane and rides the prompt (P15); ctx stays "none".
-
-_AMBIENT_TAB_KINDS = {"today", "meeting", "meetingPrep"}
-
-
-def _ambient_gated(context: "ChatContextBody | None") -> bool:
-    """Digest on/off: explicit ``include.schedule`` wins; absent → on iff the user is on a
-    meetings-relevant surface. No context (legacy client) → off (old behavior)."""
-    if context is None:
-        return False
-    include = context.include or {}
-    if isinstance(include.get("schedule"), bool):
-        return include["schedule"]
-    surface = context.surface or {}
-    if surface.get("list") == "meetings":
-        return True
-    tab = surface.get("tab") or {}
-    if tab.get("kind") in _AMBIENT_TAB_KINDS:
-        return True
-    focus = context.focus or {}
-    return focus.get("kind") == "today"
-
-
-_WORKSPACE_README_LINES = 60
-_WORKSPACE_README_CHARS = 3000
-
-
-def _fold_workspace_grounding(mounts: "list", slug: str) -> str:
-    """The workspace-focus preamble: purpose + README head for the mount matching ``slug``.
-    FAIL-CLOSED: a slug outside the caller's active/shared mounts folds nothing — the mount set
-    IS the authorization; we never read a workspace the turn couldn't see."""
-    mount = next((m for m in mounts if getattr(m, "slug", None) == slug
-                  or getattr(m, "workspace_id", None) == slug), None)
-    if mount is None:
-        return ""
-    name = str(getattr(mount, "name", "") or slug)
-    try:
-        purpose = read_purpose(mount.path) or ""
-    except Exception:  # noqa: BLE001
-        purpose = ""
-    purpose_part = f" Its purpose: {purpose.strip()}." if purpose.strip() else ""
-    readme = ""
-    try:
-        text = (Path(mount.path) / "README.md").read_text(encoding="utf-8")
-        readme = "\n".join(text.splitlines()[:_WORKSPACE_README_LINES])[:_WORKSPACE_README_CHARS]
-    except OSError:
-        readme = ""
-    fields = {"name": name, "slug": slug, "purpose": purpose_part, "readme": readme}
-    if not readme.strip():
-        return meeting_steering.NO_README_WORKSPACE_FOCUS.format(**fields)
-    return meeting_steering.render("workspace_focus", fields)
-
-
-def _enriched_meeting_focus(focus: dict, rows: "list[dict]") -> dict:
-    """Overlay the SERVER row's truth onto the client-sent meeting focus — status/title/
-    scheduled_at/workspace_id come from the meetings domain when the row is found; the client's
-    values remain only as the fallback (legacy clients / row not fetched)."""
-    nid = focus.get("native_id") or focus.get("ref")
-    row = schedule_digest_mod.find_row(
-        rows, meeting_id=focus.get("meeting_id"), platform=focus.get("platform"), native_id=nid)
-    if row is None and nid is not None:
-        # The terminal's tab param is the ROW id for planned meetings without a link (native is
-        # NULL there) — it rides in native_id, so retry it as the row id before giving up.
-        row = schedule_digest_mod.find_row(rows, meeting_id=nid)
-    if row is None:
-        return focus
-    data = row.get("data") or {}
-    merged = dict(focus)
-    merged["meeting_id"] = row.get("id", focus.get("meeting_id"))
-    merged["status"] = row.get("status") or focus.get("status")
-    if row.get("platform") and row.get("platform") != "unknown":
-        merged["platform"] = row["platform"]
-    if row.get("native_meeting_id"):
-        merged["native_id"] = row["native_meeting_id"]
-    for src_key, dst_key in (("title", "title"), ("scheduled_at", "scheduled_at"), ("workspace_id", "workspace_id")):
-        if data.get(src_key):
-            merged[dst_key] = data[src_key]
-    return merged
-
-
-def _context_grounding(
-    body: "ChatBody", session: str, redis_url: "str | None", *,
-    schedule_rows: "Callable[[], list[dict]]",
-    workspace_mounts: "Callable[[], list]",
-) -> "tuple[dict, list[str], str]":
-    """Assemble the turn's grounding from the context bundle (or the legacy ``active``).
-    ``schedule_rows`` / ``workspace_mounts`` are LAZY — fetched only for the branches that
-    need them, and both degrade to empty on failure (a bundle must never fail the turn)."""
-    prompt = body.prompt
-    context = body.context
-    focus = context.focus if context is not None else body.active
-    ctx = {"kind": "none", "session": session}
-
-    ambient = _ambient_gated(context)
-    kind = (focus or {}).get("kind")
-    need_rows = ambient or kind in ("meeting", "today")
-    rows: "list[dict]" = []
-    if need_rows:
-        try:
-            rows = schedule_rows() or []
-        except Exception:  # noqa: BLE001 — best-effort by contract
-            rows = []
-
-    tz = context.tz if context is not None else None
-    preamble = ""
-    if kind == "today":
-        digest = schedule_digest_mod.build_schedule_digest(rows, tz=tz, full_day=True)
-        if digest:
-            preamble = digest + meeting_steering.render("schedule", {})
-        return (ctx, [], preamble + prompt)
-
-    if ambient:
-        digest = schedule_digest_mod.build_schedule_digest(rows, tz=tz)
-        if digest:
-            preamble = digest + meeting_steering.render("schedule", {})
-
-    if kind == "meeting":
-        enriched = _enriched_meeting_focus(dict(focus), rows) if rows else dict(focus)
-        _c, _t, folded_prompt = _meeting_grounding(enriched, session, prompt, redis_url)
-        return (_c, _t, preamble + folded_prompt if preamble else folded_prompt)
-
-    if kind == "workspace" and (focus or {}).get("slug"):
-        try:
-            mounts = workspace_mounts() or []
-        except Exception:  # noqa: BLE001
-            mounts = []
-        preamble += _fold_workspace_grounding(mounts, str(focus["slug"]))
-        return (ctx, [], preamble + prompt)
-
-    # file focus stays client-side-preambled; none/unknown kinds fold nothing extra
-    return (ctx, [], preamble + prompt)
-
-
-# ── SSE ownership gate (P0 cross-tenant leak fix — the SSE sibling of the by-id REST check) ──────────
-# The live SSE feed `GET /api/meeting/stream` is keyed on a CALLER-SUPPLIED row id (`meeting_id`) and a
-# `session_uid`. Row ids are sequential ints, so without an ownership check any authenticated user B could
-# `EventSource(...?meeting_id=<A_row>&session_uid=<A_native>)` and stream tenant A's live transcript +
-# an ACTIVE, enumerable cross-tenant read. We mirror the WS `/ws` pattern (gateway
-# `authorize_subscribe` → `Meeting.user_id == user_id`) and the by-id REST path (`get_transcript_by_id`
-# owner-scopes in SQL): verify the caller OWNS the row BEFORE opening the redis stream. Fail CLOSED.
-#
-# agent-api has no meetings DB; it asks meeting-api `GET /meetings/{meeting_id}` forwarding the
-# gateway-injected `X-User-Id` (meeting-api's `_resolve_user_id` trusts it exactly as its by-id path does)
-# — a row owned by another user (or absent) returns 404 there → we treat it as NOT-OWNED. The returned
-# record's `native_meeting_id` also lets us confirm the requested `session_uid` belongs to the SAME owned
-# meeting, so B can't pair its own row with A's native to sniff A's feed. Returns the owned
-# meeting record (dict) on success, else None. Injectable so the L2 suite drives it over a fake.
-# How room membership was derived, stamped on every room dispatch + log line so an audit can tell
-# WHERE the room came from. Under the participant model it is the invite's addresses, resolved to
-# subjects here — the ORDER comes from the transcript, the MEMBERSHIP never does.
-_ROOM_SOURCE = "invite:participants→admin-api"
-
-
-def _http_email_subject_lookup(admin_api_url: str, internal_secret: str, admin_token: str):
-    """Build the participant ADDRESS → subject resolver: ``(address) -> str | None``.
-
-    THE DOOR PROBLEM, stated because it constrains the whole feature. agent-api already holds an
-    internal-tier seam to admin-api (``X-Internal-Secret``, used for the membership index + model
-    config), but that tier exposes NO email lookup. The one route that answers this question,
-    ``GET /admin/users/email/{email}``, is gated by ``verify_admin_token`` — a DIFFERENT and much
-    broader credential (it can also create and patch users). So:
-
-      * the narrow door is tried FIRST — ``GET /internal/users/by-email/{email}`` with the internal
-        secret. It does not exist on admin-api today; this is the route that SHOULD be added
-        (returning only ``{"id": ...}``), and when it is, the room works with no new credential and
-        ``VEXA_ADMIN_API_TOKEN`` can be dropped. A 404 here marks it absent for the process lifetime
-        so we probe once, not once per participant;
-      * the wide door is used only when an operator has explicitly set ``VEXA_ADMIN_API_TOKEN``;
-      * with neither, every lookup returns None — the room resolves to ZERO desks and says so. It
-        never falls back to matching a person by name, which is the failure this design exists to
-        avoid.
-
-    Fail-CLOSED and quiet-per-call: any error is None (that participant is skipped), never an
-    exception that could take down the turn."""
-    import urllib.error
-    import urllib.parse
-    import urllib.request
-
-    base = (admin_api_url or "").rstrip("/")
-    state = {"internal_route": True}   # flipped off the first time the narrow door 404s
-
-    def _get(url: str, headers: dict) -> "dict | None":
-        req = urllib.request.Request(url, headers=headers)
-        with urllib.request.urlopen(req, timeout=5) as resp:   # noqa: S310 — internal service URL
-            if resp.status != 200:
-                return None
-            return json.loads(resp.read().decode() or "null")
-
-    def _lookup(address: str) -> "str | None":
-        if not base or not address:
-            return None
-        quoted = urllib.parse.quote(str(address).strip().lower(), safe="")
-        if state["internal_route"] and internal_secret:
-            try:
-                row = _get(f"{base}/internal/users/by-email/{quoted}",
-                           {"X-Internal-Secret": internal_secret})
-                if isinstance(row, dict) and row.get("id") is not None:
-                    return str(row["id"])
-                return None
-            except urllib.error.HTTPError as e:
-                if e.code == 404:
-                    # Ambiguous by design: 404 is both "no such user" and "no such route". Probe the
-                    # ROUTE once — if the wide door is configured we can tell the difference by
-                    # asking it; if it is not, treat the narrow door as present and this address as
-                    # unknown (the safe reading — it skips a person rather than inventing one).
-                    if not admin_token:
-                        return None
-                    state["internal_route"] = False
-                else:
-                    return None
-            except Exception:  # noqa: BLE001 — unreachable admin-api → skip this participant
-                return None
-        if not admin_token:
-            logger.warning("room: no email→subject resolver is configured (admin-api has no "
-                           "internal by-email route and VEXA_ADMIN_API_TOKEN is unset) — the room "
-                           "will mount ZERO desks")
-            return None
-        try:
-            row = _get(f"{base}/admin/users/email/{quoted}", {"X-Admin-API-Key": admin_token})
-        except Exception:  # noqa: BLE001 — 404 (no such user) and transport errors alike → skip
-            return None
-        return str(row["id"]) if isinstance(row, dict) and row.get("id") is not None else None
-
-    return _lookup
-
-
-def _http_meeting_owner_lookup(meeting_api_url: str):
-    """Build the default owner-lookup: GET {meeting_api_url}/meetings/{id} with the caller's X-User-Id.
-    Returns a callable ``(user_id: str, meeting_id: str) -> dict | None`` (the owned meeting record, or
-    None when the row is absent / owned by someone else / meeting-api is unreachable — fail-closed)."""
-    import urllib.error
-    import urllib.request
-
-    base = (meeting_api_url or "").rstrip("/")
-
-    def _lookup(user_id: str, meeting_id: str) -> "dict | None":
-        if not base or not user_id or not str(meeting_id).isdigit():
-            return None  # non-numeric row id can't be an owned meeting row → fail closed
-        try:
-            req = urllib.request.Request(
-                f"{base}/meetings/{int(meeting_id)}", headers={"X-User-Id": str(user_id)})
-            with urllib.request.urlopen(req, timeout=5) as resp:
-                if resp.status != 200:
-                    return None
-                return json.loads(resp.read().decode() or "null")
-        except urllib.error.HTTPError:
-            return None   # 404 (not owned / absent) or any other status → refuse
-        except Exception:  # noqa: BLE001 — meeting-api unreachable → fail CLOSED, never open the stream
-            return None
-
-    return _lookup
-
+# The models, pure helpers and constants every route is built out of. They live in
+# `api_shared` so the routers can import them too — see that module's docstring.
+from control_plane.routers import health as routers_health
+from control_plane.routers import chats as routers_chats
+from control_plane.routers import admin as routers_admin
+from control_plane.routers import meetings as routers_meetings
+from control_plane.routers import scaffolds as routers_scaffolds
+from control_plane.routers import friction as routers_friction
+from control_plane.routers import workspaces as routers_workspaces
+from control_plane.api_shared import (logger, _PHASE_WORD, _iso, _provenance_line, _epoch_text, 
+    MAX_UPLOAD_BYTES, MEETING_STREAM_TRANSCRIPT_REPLAY, _upload_filename, _truncate_title, 
+    _stream_tail_id, CHAT_TURN_HEAD_TTL_SEC, _chat_turn_head_key, _record_chat_turn_head, 
+    _chat_turn_head, _Sessions, LIVE_SILENCE_TTL_SEC, _LiveMeetings, ChatContextBody, ChatBody, 
+    ScaffoldMintBody, ScaffoldHandBody, ResetBody, RoutineCreate, RoutineEnabledPatch, 
+    WorkspaceSwapBody, WorkspacePublishBody, WorkspaceRenameBody, WorkspacePushBody, 
+    GitTokenBody, WorkspacePullBody, WorkspacePurposeBody, InviteCreateBody, InviteAcceptBody, 
+    RoleSetBody, SharedNewBody, SharedAttachBody, SharedActiveBody, ArchiveBody, 
+    WorkspaceActivateBody, WorkspaceNewBody, WorkspaceDeactivateBody, _encode_sse_cursor, 
+    _decode_sse_cursor, _sse, _has_custom_model_endpoint, _model_creds_error_message, 
+    MEETING_CHAT_TRANSCRIPT_SEGMENTS, _fold_meeting_transcript, _meeting_grounding, 
+    CONTEXT_SENTINEL, _AMBIENT_TAB_KINDS, _ambient_gated, _WORKSPACE_README_LINES, 
+    _WORKSPACE_README_CHARS, _fold_workspace_grounding, _enriched_meeting_focus, 
+    _context_grounding, _ROOM_SOURCE, _http_email_subject_lookup, _http_meeting_owner_lookup)  # noqa: F401
 
 def create_app(
     dispatcher: Dispatcher,
@@ -1305,446 +377,26 @@ def create_app(
                 "group_workspace_id": group, "read_max": read_max,
                 "lookup": _email_subject_lookup}
 
-    @app.get("/health")
-    def health():
-        ok = dispatcher is not None
-        # ADDITIVE config.v1 rows (ADR-0026): the agent plane's capability tri-states (bot_gateway ·
-        # model_inference). They never affect `status`/`checks` or the status code — an unconfigured
-        # capability degrades a FEATURE (e.g. 'add bot from URL', worker model credentials), not the
-        # process; the runtime's /health carries the credentials-file probe for the mount mechanics.
-        from control_plane.config_preflight import capability_health
 
-        return JSONResponse(
-            {"status": "ok" if ok else "degraded", "service": "agent-api", "checks": {"dispatcher": ok},
-             "capabilities": capability_health()},
-            status_code=200 if ok else 503,
-        )
 
-    @app.get("/api/version")
-    def version():
-        """What is serving — unauthenticated, cheap, and polled (PRD decision 39).
 
-        No identity gate on purpose: the blue/green swap probes this from the HOST before any
-        traffic is switched onto the container, and an open browser tab polls it to notice that
-        the service underneath it moved. Both are outside a session. What it discloses is a build
-        stamp and a contract integer — the same facts the running image announces in its tag.
 
-        `api` is the pairing number the swap enforces (F55/F77): a terminal image whose baked
-        `ai.vexa.terminal.agent_api` label is not this number is REFUSED before it can be put in
-        front of a person, because that is the failure where the client leads the server and every
-        click 422s."""
-        return version_mod.version_payload()
 
-    @app.get("/api/models")
-    def models(request: Request):
-        """The ONE model this product runs: the agent's. There is no second, "streaming"/"meeting"
-        model any more — PRD decision 34 removed the in-product inference pipeline that had one."""
-        subject_of(request)  # identity gate (P20)
-        chat_model = settings.agent_model or "default"
-        return {"chat_model": chat_model, "agent_model": chat_model}
 
-    @app.post("/invocations", status_code=202)
-    def invocations(invocation: dict = Body(...)):
-        """The dispatcher sink — any trigger source POSTs a unit.v1 dispatch here."""
-        try:
-            workload_id = dispatcher.dispatch(invocation)
-        except ValidationError as e:  # non-conformant unit.v1 envelope — fail loud (P18)
-            raise HTTPException(status_code=400, detail=f"invalid unit.v1 dispatch: {e.message}")
-        return {"workload_id": workload_id}
 
-    @app.get("/api/meeting/relay-health")
-    def meeting_relay_health(request: Request):
-        """P18 (ADR 0010) — the transcript relay's observable health: is the numeric→native resolve OK,
-        and are segments arriving? A stale `VEXA_BOT_API_KEY` (401 on `/meetings`) shows here as a typed
-        `native_resolve: {ok:false, kind:'unauthorized', detail:…}` instead of silent dead air."""
-        from control_plane import transcription_watcher as _txw
-        return _txw.relay_health()
 
-    @app.get("/api/admin/overview")
-    def admin_overview(request: Request):
-        """Read-only infra + pipeline introspection for the terminal's hidden admin panel: every
-        runtime.v1 workload (agent workers + meeting bots, classified) plus the per-meeting redis
-        pipeline carriers (proc/tc streams, opt-in flag, cursor, active_meetings membership).
 
-        INTERNAL-TIER ONLY (fail-closed): the caller must present ``X-Internal-Secret`` matching
-        ``VEXA_INTERNAL_API_SECRET`` — the terminal's Next server holds it and fronts this with its
-        own email-allowlist gate; an unconfigured secret means NOBODY gets in (403), and the check
-        holds regardless of ingress (direct or via the gateway's /agent/* proxy)."""
-        from control_plane import admin_panel
 
-        secret = settings.internal_api_secret.get_secret_value() if settings is not None else ""
-        provided = request.headers.get("x-internal-secret", "")
-        if not secret or not hmac.compare_digest(provided, secret):
-            raise HTTPException(status_code=403, detail="internal secret required")
-
-        overview: dict = {"workloads": [], "meetings": []}
-        try:
-            overview["workloads"] = admin_panel.fetch_workloads(settings.runtime_api_url)
-        except Exception as e:  # noqa: BLE001 — typed partial failure (P18): the panel shows the section error
-            # SCRUBBED, like every other error this service returns (R-E11). Both of these come off
-            # a client built from a URL that routinely carries a credential — `redis://:password@host`
-            # here, an api key in the runtime URL above — and an exception's text is not ours to
-            # predict. `redact` exists two routes away and was not applied here.
-            overview["workloads_error"] = redact_secrets(f"{type(e).__name__}: {e}")
-        if redis_url:
-            import redis as _redis
-
-            try:
-                r = _redis.from_url(redis_url, decode_responses=True)
-                overview["meetings"] = admin_panel.pipeline_snapshot(r, live.list())
-            except Exception as e:  # noqa: BLE001
-                overview["meetings_error"] = redact_secrets(f"{type(e).__name__}: {e}", redis_url)
-        else:
-            overview["meetings_error"] = "no redis_url configured"
-        return overview
-
-    @app.post("/api/admin/probe")
-    def admin_probe(request: Request):
-        """Run the transcription-pipeline golden smoke probe (gateway → meeting-api → runtime →
-        redis carriers → transcript relay). Same internal-tier gate as the overview; POST because
-        it actively exercises the path (a redis write/read round-trip on scratch keys)."""
-        from control_plane import admin_panel
-        from control_plane import transcription_watcher as _txw
-
-        secret = settings.internal_api_secret.get_secret_value() if settings is not None else ""
-        provided = request.headers.get("x-internal-secret", "")
-        if not secret or not hmac.compare_digest(provided, secret):
-            raise HTTPException(status_code=403, detail="internal secret required")
-
-        r = None
-        if redis_url:
-            import redis as _redis
-
-            try:
-                r = _redis.from_url(redis_url, decode_responses=True)
-            except Exception:  # noqa: BLE001 — the probe's redis stage reports the fault
-                r = None
-        # Workloads cross-check the in-memory live registry (a stale "live" entry must not turn
-        # relay quiet into a false FAIL). Unknown (kernel unreachable) → None = trust the registry.
-        try:
-            workloads = admin_panel.fetch_workloads(settings.runtime_api_url)
-        except Exception:  # noqa: BLE001
-            workloads = None
-        return admin_panel.run_probe(settings, r, live.list(), relay_health=_txw.relay_health(),
-                                     workloads=workloads)
-
-    @app.post("/api/chat")
-    def chat(body: ChatBody, request: Request):
-        """A chat *now*-dispatch: spawn the isolated container, stream its Stream back as SSE.
-
-        RESUMABLE (mirrors /api/meeting/stream): every SSE event carries an ``id:`` = the unit output
-        Stream cursor. A dropped view (per-dispatch worker cold-start races the SSE, a transient proxy
-        drop) reconnects with ``Last-Event-ID`` — we then RE-ATTACH to the SAME warm unit and resume the
-        read from that cursor (gapless) WITHOUT dispatching a second turn. The turn was never lost (the
-        worker completes + commits regardless); resume just re-shows the output the client missed."""
-        if stream_reader is None:
-            raise HTTPException(status_code=501, detail="stream relay not wired")
-        subject = subject_of(request)  # server-derived (P20); body.subject is ignored
-        session = body.session or units.DEFAULT_CHAT_SESSION
-        # THE MEETING ROOM (post-meeting run). Resolved and AUTHORISED here, before anything else
-        # happens on this request — a refusal must cost the caller a 403, never a partially-run turn.
-        # ``room`` never comes from the body: the body names a MEETING (and may PROPOSE a narrowing);
-        # every subject in the returned room came from meeting-api. See ``_resolve_room``.
-        room = None
-        if body.room_meeting_id:
-            room = _resolve_room(request, subject, body.room_meeting_id,
-                                 participants=body.room_participants,
-                                 names=body.room_participant_names,
-                                 speakers=body.room_speakers,
-                                 read_max=body.room_read_max)
-        # THE SCAFFOLD (PRD 5.5). When the turn names one and it is THIS subject's, the record
-        # — not the client — decides two things: which workspaces this chat mounts, and what
-        # the opening ask is. Both used to be composed client-side, and that is precisely why
-        # the panel and the agent disagreed about the same click. A scaffold that is not
-        # this subject's is IGNORED rather than refused: a forwarded or stale id must not be
-        # able to widen anybody's mounts, and must not break their turn either.
-        scaffold_view = None
-        if body.scaffold_id:
-            rec = scaffolds.get(body.scaffold_id)
-            if rec is not None and _scaffold_is_for(rec, request, subject):
-                rec = scaffolds.redeem(body.scaffold_id, subject) or rec
-                try:
-                    scaffold_view = _scaffold_view(rec, subject)
-                except scaffolds_mod.ScaffoldError:
-                    logger.warning("scaffold %s cannot be rendered (its preset is gone) — the "
-                                   "turn runs without it", body.scaffold_id)
-            else:
-                logger.warning("scaffold %s ignored on a turn by subject=%s (unknown or not "
-                               "theirs)", body.scaffold_id, subject)
-        if scaffold_view is not None:
-            # THE OPENING IS THE RECORD'S, SUBSTITUTED HERE. The terminal composes no text —
-            # it sends the id. The facts ride in front of the ask so the agent's first turn
-            # can name the meeting, the time, the room and the person's state without
-            # fetching anything, and the whole block is machinery-marked so the human never
-            # sees it as their own message (ledger F7).
-            body = body.model_copy(update={"prompt": scaffolds_mod.turn_prompt(scaffold_view)})
-        # THE INTENT'S PRESET (decision 32.2 / 35.3). Same rule as the scaffold above and for the
-        # same reason: the words are admin-owned content in `_global/asks/`, the wire carries a kind
-        # and its arguments, and the server is the only thing that puts the two together.
-        #
-        # DEGRADES, NEVER REFUSES. `preset_for` returns None for a kind this deployment does not
-        # know and `read_preset` raises when the file is not there; both leave `body.prompt` — the
-        # client's plain fallback sentence — exactly as it arrived. A preset library that is one
-        # release behind the terminal costs the turn its phrasing, not its meaning.
-        elif body.intent:
-            _preset = chat_intents.preset_for(body.intent)
-            if _preset:
-                try:
-                    _fm, _ask = scaffolds_mod.read_preset(_global_root(), _preset)
-                    _text = scaffolds_mod.substitute(_ask, chat_intents.tokens_for(body.intent))
-                    # A SILENT KIND IS MACHINERY END TO END (decision 35). The marks ride the prompt
-                    # itself — the same carrier the write-back phase uses — so `workspace_reader.
-                    # history` drops this turn and every agent turn after it until the person speaks
-                    # again. Nothing downstream needs a new field, and a deployment whose reader is
-                    # older simply renders a marked turn it does not yet hide, rather than breaking.
-                    if chat_intents.is_silent(body.intent):
-                        _text = chat_intents.SILENT_PREFIX + _text
-                    body = body.model_copy(update={"prompt": _text})
-                except scaffolds_mod.ScaffoldError as e:
-                    logger.warning("intent %s has no preset here (%s) — the turn runs on the "
-                                   "client's fallback sentence", _preset, e)
-        # A reconnect carries Last-Event-ID (the last Stream cursor the client rendered). On resume we
-        # DON'T re-dispatch — we re-attach to the existing warm unit and read from the cursor onward.
-        resume = request.headers.get("last-event-id") or None
-        # Ground the chat in the terminal's ACTIVE meeting (if any): agent-api folds the live transcript
-        # from the meeting's redis Stream (tc:meeting:{native} — the SAME stream the live view renders) into
-        # the prompt, fresh on every turn. The transcript stays inside the trusted control plane and
-        # rides the prompt to the worker — no file, no cross-domain HTTP, no user key in the worker (P15).
-        ctx, tools, prompt = _context_grounding(
-            body, session, redis_url,
-            schedule_rows=lambda: _schedule_source(subject),
-            workspace_mounts=lambda: (active_workspaces(wsr.root, subject)
-                                      + shared_active_mounts(wsr.root, subject, mindex.list(subject))),
-        )
-        # Mark the grounding→user boundary. Every branch returns `<grounding> + body.prompt`, so the
-        # user's words are the exact suffix; the sentinel goes right before them.
-        #
-        # ⚠ IT USED TO SKIP THE TURNS THAT NEEDED IT MOST. The condition carried `len(prompt) >
-        # len(body.prompt)` — "only mark it when I actually folded something" — which is wrong twice
-        # over: the WORKER prepends its own preambles (voice, kg-links, mount stack, entity index,
-        # global context) AFTER this function returns, so a turn this function folded nothing into
-        # still reaches the transcript with several screens of machinery in front of the sentence.
-        # That is the exact shape of the 2026-09-02 regression: the founder's turns had no meeting,
-        # no schedule and no workspace grounding, so no sentinel was written, so the terminal fell
-        # through to its regexes, which no longer matched the preambles — and his whole machinery
-        # prompt rendered as a grey USER bubble. A boundary marker that is present only sometimes is
-        # a boundary marker nobody can rely on. Now: any turn carrying the person's words carries it.
-        if body.prompt and prompt.endswith(body.prompt):
-            prompt = prompt[: len(prompt) - len(body.prompt)] + CONTEXT_SENTINEL + body.prompt
-        # Attribute this turn's commits to the human editor by EMAIL (gateway-injected, trusted) rather
-        # than the bare subject id — the git author NAME becomes the email; the synthetic author email
-        # (<subject>@vexa.local) stays for the you/member classification (workspace_reader.git_state_at).
-        _email = (request.headers.get("x-user-email") or "").strip()
-        inv = units.make_dispatch(
-            subject=subject, trigger="message",
-            start=units.entrypoint(inline=prompt), context=ctx, tools=tools,
-            principal={"name": _email} if _email else None,
-        )
-        if resume:
-            # Re-attach only — the warm unit id is deterministic from (subject, session); resume reads
-            # its durable output Stream from the cursor. No new turn, no session re-title.
-            unit_id = units.dispatch_id(inv)
-        else:
-            unit_id = units.dispatch_id(inv)
-            retry_from = _chat_turn_head(redis_url, unit_id, body.turn_id) if body.turn_id else None
-            if retry_from is not None:
-                # No-cursor RETRY of the current turn (the stream dropped before the client saw any
-                # ``id:``): re-attach from the turn's recorded start — the whole turn replays, including
-                # a terminal event the worker wrote while the client was gone. NO second dispatch.
-                resume = retry_from
-            else:
-                # Fresh turn — credential preflight FIRST (config.v1 ``model_inference``, the
-                # request-path oracle): with no deployment credential AND no per-user custom
-                # endpoint, the worker's claude CLI can only fail with its own "Not logged in ·
-                # Please run /login" — an adapter internal that means nothing to an API consumer.
-                # Refuse HERE with an actionable frame instead: no worker spawn, no ghost session
-                # entry. A FAILED config lookup (None) fails OPEN — a down identity service must
-                # never block a turn; the worker-side auth taxonomy still catches it cleanly.
-                if capability_state("model_inference") == NOT_CONFIGURED:
-                    cfg = dispatcher.resolve_model_config(subject)
-                    if cfg is not None and not _has_custom_model_endpoint(cfg):
-                        return StreamingResponse(
-                            _sse([{"type": "error", "message": _model_creds_error_message()},
-                                  {"type": "turn-complete"}]),
-                            media_type="text/event-stream",
-                            headers={"Cache-Control": "no-cache", "X-Accel-Buffering": "no",
-                                     "X-Unit-Id": unit_id, "X-Chat-Session": session},
-                        )
-                # Snapshot the out-Stream tail BEFORE dispatching and attach the reader from
-                # it — attaching at ``$`` raced the worker (events written between dispatch and attach,
-                # or a whole turn that finished in the gap, were invisible: the 'Reconnecting' hang).
-                # The thread's Stream holds PRIOR turns too, so the snapshot (not stream start) is the
-                # earliest safe attach point — the client appends and stops on any ``turn-complete``.
-                start = _stream_tail_id(redis_url, units.output_topic(unit_id)) or None
-                # Upsert the durable index on first use of a thread: a new thread is titled by its first
-                # prompt; an existing one just bumps last_active (title preserved).
-                is_new = not any(r["session"] == session for r in sess.list(subject))
-                # A SCAFFOLDED chat is titled by its record, never by its opening: the opening
-                # is machinery, and titling a rail row with the first 60 characters of an
-                # instruction block is the same defect as painting it as the person's message.
-                _title = ((scaffold_view["header"]["title"] or scaffold_view["opening_label"])
-                          if scaffold_view is not None else _truncate_title(body.prompt))
-                sess.upsert(subject, session, title=_title if is_new else None)
-                # ``room`` applies AT SPAWN: the mount table is fixed when the container is created,
-                # so a WARM unit (this thread already has a live worker) keeps the stack it booted
-                # with and a room named on a later turn of the same thread does not retro-mount. The
-                # post-meeting run uses its own per-meeting session, so it spawns cold and gets the
-                # room; a turn that needs a different room needs a different session.
-                try:
-                    unit_id = dispatcher.dispatch(  # spawn-or-touch the thread's warm chat unit
-                        inv, room=room,
-                        scaffold_workspaces=(scaffold_view or {}).get("workspaces") or None)
-                except dispatch_mod.WarmDeliveryFailed as exc:
-                    # THE TURN IS REFUSED, NOT DROPPED. For a warm unit the pre-delivery is the only
-                    # delivery, so a failure here means the person's words reached nobody. Answering
-                    # 200 and streaming the turn already in flight is what made this invisible: they
-                    # watch a reply appear and reasonably believe it is to what they just sent.
-                    # 503 is deliberate — this is transient by nature (redis blip, a worker in the
-                    # idle-exit race) and the honest instruction is "send it again".
-                    logger.warning("chat turn refused for subject=%s session=%s: %s",
-                                   subject, session, exc)
-                    raise HTTPException(
-                        status_code=503,
-                        detail="That message did not reach your agent — nothing was lost on your "
-                               "side, please send it again.") from exc
-                if body.turn_id and start is not None:
-                    _record_chat_turn_head(redis_url, unit_id, body.turn_id, start)
-                resume = start
-        return StreamingResponse(
-            _sse(stream_reader.read(unit_id, resume=resume)),
-            media_type="text/event-stream",
-            headers={"Cache-Control": "no-cache", "X-Accel-Buffering": "no",
-                     "X-Unit-Id": unit_id, "X-Chat-Session": session},
-        )
-
-    @app.post("/api/chat/reset")
-    def chat_reset(body: ResetBody, request: Request):
-        """Drop a conversation thread: remove it from the index AND delete its continuity file so a
-        future turn on the same name starts a fresh conversation (not a resume of the old one)."""
-        subject = subject_of(request)
-        session = body.session or units.DEFAULT_CHAT_SESSION
-        sess.drop(subject, session)
-        try:
-            wsr.drop_session(subject, session)
-        except Exception:  # noqa: BLE001 — index drop is the contract; the file delete is best-effort
-            logger.exception("dropping continuity file failed subject=%s session=%s", subject, session)
-        return {"ok": True}
-
-    @app.get("/api/sessions")
-    def list_sessions(request: Request):
-        return {"sessions": sess.list(subject_of(request))}
-
-    @app.get("/api/sessions/{session}/history")
-    def session_history(session: str, request: Request):
-        """The session's prior conversation, as simplified turns the terminal can render (so clicking a
-        saved chat re-opens its history). Tolerant: a missing/empty transcript returns ``{turns: []}``;
-        an invalid subject/session never 500s."""
-        subject = subject_of(request)
-        # The turn's cwd FOLLOWS the active set (flat model), so a thread's continuity may sit under
-        # any currently-mounted workspace dir — hand the reader those candidates. Best-effort: a
-        # failing mount resolution only narrows the search to _system + home.
-        extra: list = []
-        try:
-            ms = active_workspaces(wsr.root, subject) + shared_active_mounts(wsr.root, subject, mindex.list(subject))
-            extra = [m.path for m in ms]
-        except Exception:  # noqa: BLE001
-            logger.warning("mount resolution for history failed subject=%s — searching anchored roots only", subject)
-        try:
-            turns = wsr.history(subject, session, extra_roots=extra)
-        except Exception:  # noqa: BLE001 — history is best-effort; a bad path → empty, never an error
-            logger.exception("loading session history failed subject=%s session=%s", subject, session)
-            turns = []
-        return {"turns": turns}
 
     # ── routines (MVP2) — a scheduled routine compiles to a schedule.v1 cron job whose body is a
     #    unit.v1 dispatch POSTed back to /invocations when due (the runtime owns the durable cron) ──
-    @app.post("/api/routines", status_code=201)
-    def create_routine(body: RoutineCreate, request: Request):
-        if scheduler is None or not invocations_url:
-            raise HTTPException(status_code=501, detail="scheduler not wired")
-        try:
-            routine = routines_mod.make_routine(
-                subject=subject_of(request), name=body.name, cron=body.cron, prompt=body.prompt,
-            )
-            job_spec = routines_mod.compile_to_job(routine, invocations_url=invocations_url)
-        except (ValueError, ValidationError) as e:  # bad cron form / non-conformant routine — fail loud
-            raise HTTPException(status_code=400, detail=str(getattr(e, "message", e)))
-        job = scheduler.schedule(job_spec)
-        ran_now = False
-        if body.run_now:
-            # Fire one immediate run via the dispatcher (no HTTP hop) so the author sees a result now.
-            try:
-                dispatcher.dispatch(job_spec["request"]["body"])
-                ran_now = True
-            except Exception:  # noqa: BLE001 — the routine is still scheduled even if the demo run fails
-                ran_now = False
-        return {"routine": routine, "job_id": job.get("job_id"), "ran_now": ran_now}
 
-    @app.get("/api/routines")
-    def list_routines(request: Request):
-        if scheduler is None:
-            return {"routines": []}
-        cards = workspace_routines_mod.routine_cards_for_subject(
-            subject_of(request),
-            jobs=scheduler.list_jobs(limit=1000),
-            workspaces_dir=wsr.root,
-        )
-        return {"routines": cards}
 
-    @app.patch("/api/routines/{name}/enabled")
-    def set_routine_enabled(name: str, body: RoutineEnabledPatch, request: Request):
-        if scheduler is None or not invocations_url:
-            raise HTTPException(status_code=501, detail="scheduler not wired")
-        subject = subject_of(request)
-        try:
-            workspace_routines_mod.set_routine_file_enabled(
-                subject,
-                name,
-                enabled=body.enabled,
-                workspaces_dir=wsr.root,
-            )
-            result = workspace_routines_mod.reconcile_workspace_routines(
-                subject,
-                scheduler=scheduler,
-                invocations_url=invocations_url,
-                workspaces_dir=wsr.root,
-            )
-        except FileNotFoundError:
-            raise HTTPException(status_code=404, detail="unknown routine")
-        except ValueError as e:
-            raise HTTPException(status_code=400, detail=str(e))
-        return {
-            "ok": True,
-            "name": name,
-            "enabled": body.enabled,
-            "reconcile": result.__dict__,
-        }
 
-    @app.delete("/api/routines/{routine_id}")
-    def delete_routine(routine_id: str, request: Request):
-        if scheduler is None:
-            raise HTTPException(status_code=501, detail="scheduler not wired")
-        subject = subject_of(request)
-        for job in scheduler.list_jobs():
-            meta = job.get("metadata") or {}
-            if meta.get("routine_id") == routine_id and meta.get("owner") == subject:
-                scheduler.cancel_job(job["job_id"])
-                return {"ok": True, "routine_id": routine_id}
-        raise HTTPException(status_code=404, detail="unknown routine")
 
     # ── events (MVP3) — the GENERIC event-source ingress: any event.v1 Event → a unit.v1 dispatch →
     #    the one Dispatcher. agent-api knows no tool/domain; the unit reaches email/calendar via its
     #    toolbelt. Email-triage, post-meeting, news all POST here (one front door, P6) ──
-    @app.post("/events", status_code=202)
-    def events(event: dict = Body(...)):
-        try:
-            invocation = event_to_invocation(event)
-        except ValidationError as e:
-            raise HTTPException(status_code=400, detail=f"invalid event.v1: {e.message}")
-        except ValueError as e:  # no plan carried — fail loud (P18)
-            raise HTTPException(status_code=422, detail=str(e))
-        workload_id = dispatcher.dispatch(invocation)
-        return {"workload_id": workload_id, "trigger": invocation["trigger"]}
 
     def _read_target(request: Request, slug: Optional[str], *, write: bool = False) -> Path:
         """Resolve which workspace dir a read or write targets, returning its ABSOLUTE PATH. Default (no
@@ -2071,266 +723,10 @@ def create_app(
             "share_token": str(share_token) if share_token else None,
         })
 
-    @app.post("/internal/scaffolds", status_code=201)
-    def mint_scaffold(body: ScaffoldMintBody, request: Request):
-        """Mint one scaffold and return `{id, url}` — the INTERNAL tier only (flows and the rig).
 
-        Gated on `X-Internal-Secret`, the same edge the meeting room uses, for the same reason: a
-        scaffold names the workspaces a chat will mount and the opening its agent will answer, so a
-        caller who could mint one for somebody else's address could compose that person's first
-        turn. A browser client through the gateway holds no such secret and cannot mint at all.
 
-        EVERYTHING THAT CAN BE WRONG IS WRONG HERE, NOT AT THE CLICK. The preset must exist and be
-        non-empty, the kind must be in the catalogue, the terminal's origin must be configured. A
-        step calls this BEFORE it sends, so a failure here stops the send — which is the share-gate
-        doctrine one layer up (`email_attendees` refuses to mail a link it cannot make work). A
-        record that mints happily and opens onto nothing is the exact failure this route exists to
-        make impossible.
 
-        THE SHARE TOKEN IS MINTED BY THE CALLER, not here, and that is a deliberate boundary. The
-        restricted grant is `POST /meetings/{id}/share` on meeting-api as the meeting's OWNER, which
-        needs that owner's gateway key; flows already holds the credential that can produce one
-        (`flows_steps/meeting.mint_transcript_share`) and agent-api holds no gateway key by design.
-        Giving agent-api one so this route could mint the share would put a key that can act as any
-        user into a service whose whole job is to read workspaces. The caller mints, and passes the
-        token here; this route composes it into the url so the LINK is still built in one place."""
-        if not _internal_caller(request):
-            raise HTTPException(status_code=403, detail="minting a scaffold is an internal-tier capability")
-        who = str(body.who or "").strip()
-        if not who or "@" not in who:
-            raise HTTPException(status_code=400, detail="`who` must be the recipient's address")
-        if body.kind not in scaffolds_mod.KINDS:
-            raise HTTPException(status_code=400,
-                                detail=f"unknown scaffold kind {body.kind!r} — the catalogue is "
-                                       f"{', '.join(scaffolds_mod.KINDS)}")
-        ui = (settings.ui_url if settings is not None else "").rstrip("/")
-        if not ui:
-            raise HTTPException(status_code=503,
-                                detail="VEXA_UI_URL is not set on agent-api — a scaffold url would "
-                                       "have no origin, and a link nobody can open is not a touch")
-        try:
-            fm, _body_text = scaffolds_mod.read_preset(_global_root(), str(body.opening or ""))
-        except scaffolds_mod.ScaffoldError as e:
-            raise HTTPException(status_code=400, detail=str(e)) from e
 
-        subject = _email_subject_lookup(who.lower()) or ""
-        row = None
-        mid = str(body.meeting or "").strip()
-        minted_by = str((body.provenance or {}).get("minted_by") or "")
-        if mid.isdigit():
-            for reader_uid in (subject, minted_by):
-                if reader_uid:
-                    row = _meeting_owner_lookup(reader_uid, mid)
-                    if isinstance(row, dict):
-                        break
-        group = scaffolds_mod.group_workspace_of(row)
-
-        # THE MOUNT SET, STATED. `_global` always (it is the company layer and every worker carries
-        # it), the recipient's own desk when they have one, the group desk when the meeting is bound
-        # to one. A caller may override the whole list; it may not omit `_global`.
-        if body.workspaces is not None:
-            mounts = [w for w in body.workspaces if str(w).strip()]
-        else:
-            mounts = [subject] if subject else []
-            if group and group not in mounts:
-                mounts.append(group)
-        mounts = [system_mounts.GLOBAL_SLUG] + [m for m in mounts if m != system_mounts.GLOBAL_SLUG]
-
-        # ONE FIRST VISIT PER PERSON, NOT ONE PER SIGN-IN.
-        #
-        # Every other kind is minted by a flow at the moment it creates a touch, so a second mint is
-        # a second touch and deserves its own record. `first-visit` is the opposite: it is minted
-        # because nobody sent anything, on every sign-in that arrives without a link. Minting a
-        # fresh record each time would give the terminal a new `scaffold-<id>` chat id each time,
-        # and therefore a new rail row — a fourth sign-in showing four rows the person never made,
-        # which is precisely the defect F34/F35 exist to remove. Caught in review by the terminal
-        # worker before it shipped.
-        #
-        # So the record is reused while it is still PENDING. Once redeemed the person has actually
-        # been through it, and a later sign-in is a genuinely new arrival that may hold new facts —
-        # a workspace shared with them since, a meeting they have been invited to.
-        if body.kind == "first-visit":
-            existing = next((r for r in scaffolds.for_recipient(who, pending_only=True)
-                             if r.get("kind") == "first-visit"), None)
-            if existing:
-                logger.info("scaffold REUSED id=%s kind=first-visit who=%s (unredeemed)",
-                            existing["id"], who)
-                return {"id": existing["id"], "url": f"{ui}/?s={existing['id']}"}
-
-        rec = _compose_and_mint(
-            who=who, subject=subject, kind=body.kind, opening=str(body.opening), fm=fm,
-            mid=mid, mounts=mounts, group=group, refs_in=body.refs,
-            tabs=body.tabs, focus=body.focus, provenance=body.provenance,
-            share_token=body.share_token,
-        )
-        # THE LINK IS AN ID. Nothing else: no preset name, no mount list, no prompt text — and since
-        # R-A08, no capability either. The share the caller minted is stored ON the record and the
-        # recipient redeems it against this id (`POST /api/scaffolds/<id>/share`), because a bearer
-        # token in a query string leaks into every access log and proxy trace it passes through —
-        # the rule `worker/engine.py` states for the delegation token, applied to the artefact that
-        # actually crosses a mail provider.
-        url = f"{ui}/?s={rec['id']}"
-        logger.info("scaffold MINTED id=%s kind=%s who=%s meeting=%s mounts=%s opening=%s share=%s",
-                    rec["id"], rec["kind"], who, rec.get("meeting"), mounts, rec["opening"],
-                    bool(body.share_token))
-        return {"id": rec["id"], "url": url}
-
-    @app.post("/api/scaffolds/hand", status_code=201)
-    def mint_hand_scaffold(body: ScaffoldHandBody, request: Request):
-        """A hand link becomes a record, minted FOR THE CALLER, and the client composes nothing.
-
-        `/?ask=<preset>&meeting=<row>` is the one composition path the scaffold work left standing,
-        and it was still being composed in the browser: the terminal read the preset and substituted
-        `?meeting=` and `?ws=` into the opening text. A URL that reaches the model is the thing
-        decision 13 forbids — the URL carries NAMES, never prompt text — and the `?s=` path obeyed
-        it while this one did not. So the hand link mints here and redirects to `/?s=<id>`, and the
-        substitution happens where every other scaffold's does: server-side, at turn time, out of
-        the RECORD (`scaffolds.turn_prompt`), never out of the address bar.
-
-        THREE THINGS THIS ROUTE REFUSES, and each is the whole point of it existing:
-
-          · a preset that is not a NAME in `_global/asks/` — `read_preset` decides, exactly as the
-            internal mint does, so there is one definition of what an opening may be;
-          · a `who` other than the caller — there is no such field. The recipient is the session's
-            own subject and address, so a link cannot mint a first turn for somebody else;
-          · a meeting the caller cannot see. `_meeting_owner_lookup` is owner-only and fail-closed,
-            so a crafted `?meeting=<someone else's row>` mints NOTHING rather than a record carrying
-            another meeting's title and attendees into this person's opening.
-
-        The last one is the subtle half. Even with no text in the URL, an unchecked row id would let
-        a crafted link decide WHICH FACTS the agent is handed — which is the same attack one level
-        down. A hand link may name only a meeting you own; an attendee reaches theirs through the
-        emailed scaffold, which carries its own share."""
-        subject = subject_of(request)
-        if not subject:
-            raise HTTPException(status_code=401, detail="sign in first")
-        # The caller's ADDRESS, gateway-injected and trusted — the same header the chat turn already
-        # attributes commits by, and the gateway strips any client-supplied spelling of it before it
-        # gets here (see the TOPOLOGY BOUNDARY note). A record is bound to an address, so a session
-        # that carries none cannot mint: refusing is honest, guessing one would bind the record to a
-        # person who may not be the caller.
-        who = (request.headers.get("x-user-email") or "").strip()
-        if not who or "@" not in who:
-            raise HTTPException(status_code=409,
-                                detail="this session carries no address, and a scaffold is bound to "
-                                       "one — sign in again")
-        preset = str(body.preset or "").strip()
-        if not scaffolds_mod.NAME_RE.match(preset):
-            raise HTTPException(status_code=400, detail="`preset` must be a preset name")
-        try:
-            fm, _text = scaffolds_mod.read_preset(_global_root(), preset)
-        except scaffolds_mod.ScaffoldError as e:
-            raise HTTPException(status_code=400, detail=str(e)) from e
-
-        ui = (settings.ui_url if settings is not None else "").rstrip("/")
-        if not ui:
-            raise HTTPException(status_code=503, detail="VEXA_UI_URL is not set on agent-api")
-
-        mid = str(body.meeting or "").strip()
-        row = None
-        if mid:
-            row = _meeting_owner_lookup(subject, mid) if mid.isdigit() else None
-            if not isinstance(row, dict):
-                # fail CLOSED and say so: a link naming a meeting this person cannot open is a
-                # broken link, and minting a record without it would answer about the wrong thing.
-                raise HTTPException(status_code=404,
-                                    detail="no such meeting for this account — a hand link may only "
-                                           "name a meeting you can open")
-        group = scaffolds_mod.group_workspace_of(row)
-        mounts = [system_mounts.GLOBAL_SLUG, subject] + ([group] if group else [])
-
-        rec = _compose_and_mint(
-            who=who, subject=subject, kind="hand-link", opening=preset, fm=fm,
-            mid=mid, mounts=mounts, group=group, refs_in=None,
-            tabs=None, focus=None,
-            provenance={"flow": "hand-link", "minted_by": subject},
-        )
-        logger.info("scaffold MINTED (hand) id=%s who=%s meeting=%s opening=%s",
-                    rec["id"], who, rec.get("meeting"), rec["opening"])
-        return {"id": rec["id"], "url": f"{ui}/?s={rec['id']}"}
-
-    @app.get("/api/scaffolds/{scaffold_id}")
-    def read_scaffold(scaffold_id: str, request: Request):
-        """The record, as its recipient. Refuses anyone else; marks REDEEMED on first read.
-
-        A 404 for an unknown id AND for one that is not yours — the id is a capability until redeem
-        binds it, so a 403 would tell a prober that a scaffold with that id exists."""
-        subject = subject_of(request)
-        rec = scaffolds.get(scaffold_id)
-        if rec is None:
-            raise HTTPException(status_code=404, detail="no such scaffold")
-        if not (_internal_caller(request) or _scaffold_is_for(rec, request, subject)):
-            logger.warning("scaffold REFUSED id=%s subject=%s reason=not-the-recipient", scaffold_id, subject)
-            raise HTTPException(status_code=404, detail="no such scaffold")
-        # REDEEM ONLY FOR THE RECIPIENT (R-A13). An admin's debugging read used to stamp
-        # `redeemed_at`/`redeemed_by` with the ADMIN — and `scaffolds.redeem`'s own docstring calls
-        # that stamp "the only measurement the alpha ledger's 'seconds to act' column is made of".
-        #
-        # It is load-bearing twice over, which is why it is fixed here rather than left on the
-        # backlog: `redeemed_by == subject` is one of the three identity tests BOTH scaffold
-        # predicates run, so a stamp written by an admin's read promoted that admin to "the
-        # recipient" on every later call — including the share hand-out one route down, which is
-        # exactly the capability R-A08 exists to keep bound to one person.
-        if _scaffold_recipient_is(rec, request, subject):
-            rec = scaffolds.redeem(scaffold_id, subject) or rec
-        try:
-            return _scaffold_view(rec, subject)
-        except scaffolds_mod.ScaffoldError as e:
-            # The preset was there at mint and is not there now — an admin deleted or emptied it.
-            # Say so; a silent empty opening is the failure that let the phase greeting win (F5).
-            raise HTTPException(status_code=409, detail=str(e)) from e
-
-    @app.post("/api/scaffolds/{scaffold_id}/share")
-    def redeem_scaffold_share(scaffold_id: str, request: Request):
-        """The transcript share this record carries, handed to its RECIPIENT — the replacement for
-        `&tshare=` in the mailed link (R-A08).
-
-        A scaffold for a meeting that is not the reader's own carries a restricted grant, minted by
-        the meeting's owner in `core/flows` and passed to the mint. It used to ride the link's query
-        string, where a bearer credential enters every access log and proxy trace between us and the
-        recipient's inbox — and then whatever they forward. `worker/engine.py:1044-1046` states the
-        rule for the delegation token in as many words; this was the weaker spelling on the more
-        exposed artefact.
-
-        `{"token": null}` — not a 404 — when the record carries no share: most scaffolds are about
-        the reader's own meeting and carry no capability, and answering "not found" for the ordinary
-        case would teach the client to treat it as breakage.
-
-        A 404 for a stranger AND for an unknown id, exactly as the read route answers, so this route
-        cannot be used to discover that a scaffold id exists. An INSTANCE ADMIN gets the 404 too —
-        see `_scaffold_recipient_is` — and so does the internal tier: the mint hands the token IN,
-        and a route that handed it back out would be a second way to reach it."""
-        subject = subject_of(request)
-        rec = scaffolds.get(scaffold_id)
-        if rec is None or not _scaffold_recipient_is(rec, request, subject):
-            logger.warning("scaffold share REFUSED id=%s subject=%s reason=not-the-recipient",
-                           scaffold_id, subject)
-            raise HTTPException(status_code=404, detail="no such scaffold")
-        rec = scaffolds.hand_share(scaffold_id, subject) or rec
-        token = rec.get("share_token") or None
-        if token:
-            logger.info("scaffold share HANDED id=%s subject=%s", scaffold_id, subject)
-        return {"token": token}
-
-    @app.get("/api/scaffolds")
-    def list_scaffolds(request: Request, mine: bool = True, pending: bool = True):
-        """The caller's own scaffolds — what is WAITING for this person behind a link they may never
-        have clicked. Step 6 of the build order (`whats_waiting` returns pending scaffolds so the
-        person's own agent opens the same record) reads exactly this: one record, two renderers."""
-        subject = subject_of(request)
-        email = (request.headers.get("x-user-email") or "").strip().lower()
-        # Indexed by ADDRESS, because that is what a scaffold is minted against — the recipient
-        # usually has no subject yet. With no gateway-injected address there is nothing to look up,
-        # and the honest answer is an empty list, never every scaffold on the instance.
-        rows = scaffolds.for_recipient(email, pending_only=pending) if email else []
-        out = []
-        for r in rows:
-            try:
-                out.append(_scaffold_view(r, subject))
-            except scaffolds_mod.ScaffoldError:
-                # One scaffold whose preset an admin deleted must not empty the whole list.
-                logger.warning("scaffold %s is unrenderable (its preset is gone) — omitted", r.get("id"))
-        return {"scaffolds": out}
 
     # ── ROUGH EDGES (PRD decision 33) ───────────────────────────────────────────────────────────
     def _friction_subject(request: Request) -> str:
@@ -2348,163 +744,12 @@ def create_app(
         into ONE row with a counter."""
         return (request.headers.get("x-user-id") or "").strip()
 
-    @app.post("/api/friction", status_code=201)
-    def file_friction(body: dict, request: Request):
-        """File one rough edge — from an agent (`report_friction`), the harness, or a person's
-        "Report this" (decision 33 §§1–2). Returns the stored record; `recurrence > 1` means this
-        edge was already known and this report is another occurrence of it."""
-        # THE SUBJECT IS THE CALLER'S, NEVER THE BODY'S (R-E04). The route is deliberately
-        # unauthenticated — an unattributable report beats none — but a body that could name a
-        # subject let an anonymous caller write records attributed to a named user, which is worse
-        # than an unattributed record in exactly the way a forged signature is worse than none.
-        rec = friction.file({**(body or {}), "subject": _friction_subject(request)})
-        logger.info("friction filed id=%s kind=%s status=%s recurrence=%s tool=%s",
-                    rec.get("id"), rec.get("kind"), rec.get("status"), rec.get("recurrence"),
-                    (rec.get("context") or {}).get("tool") or "-")
-        return {"id": rec["id"], "status": rec["status"], "recurrence": rec["recurrence"],
-                "kind": rec["kind"], "known": rec["recurrence"] > 1}
 
-    @app.get("/api/friction/dump")
-    def dump_friction(request: Request, since: str = "", status: str = "open",
-                      format: str = "md"):
-        """THE DUMP (decision 33 §3) — the open rough edges as a brief a fixing agent works off.
 
-        `format=md` (default) returns the markdown; `format=json` returns the same grouping as data,
-        which is what the rig's `friction_so_far` renders. Both come from ONE renderer in
-        `shared/friction.py`: a dump and a tool that disagree about what is open is the two-renderers
-        failure this codebase has already paid for twice."""
-        subject = subject_of(request)   # a read: identified callers only, like every other read here
-        ts = friction_store_mod.parse_since(since)
-        rows = friction.since(ts, status=status)
-        # A report carries the workspace names, paths and free text of what a person was DOING when
-        # it broke. So the dump is the caller's own, with one exception: the fixing agent of
-        # decision 33 §3 is the instance admin, and it is the whole point of the dump (R-E05).
-        if not global_layer.is_admin(settings, str(subject)):
-            rows = [r for r in rows if str(r.get("subject") or "") == str(subject)]
-        if format == "json":
-            return {"since": since, "status": status, "count": len(rows),
-                    "findings": friction_mod.group(rows), "records": rows}
-        return Response(content=friction_mod.render_markdown(rows, since=since, status=status),
-                        media_type="text/markdown; charset=utf-8")
 
-    @app.post("/api/friction/{friction_id}/fix")
-    def fix_friction(friction_id: str, body: dict, request: Request):
-        """Close one record against the change that addressed it (decision 33 §4).
 
-        A record filed again after this flips itself to `recurring` — which is why closing is cheap
-        and why a fixing agent should close what it addressed rather than waiting to be sure."""
-        subject = subject_of(request)
-        # Same scoping as the dump, and a 404 rather than a 403 for someone else's record: the id is
-        # the only thing a prober would learn from the difference (R-E05).
-        existing = friction.get(friction_id)
-        if existing is None or not (global_layer.is_admin(settings, str(subject))
-                                    or str(existing.get("subject") or "") == str(subject)):
-            raise HTTPException(status_code=404, detail="no such friction record")
-        try:
-            rec = friction.fix(friction_id, str((body or {}).get("fix_ref") or ""))
-        except ValueError as e:
-            raise HTTPException(status_code=400, detail=str(e)) from e
-        if rec is None:
-            raise HTTPException(status_code=404, detail="no such friction record")
-        return {"id": rec["id"], "status": rec["status"], "fix_ref": rec["fix_ref"]}
 
-    @app.get("/api/workspace/tree")
-    def ws_tree(request: Request, hidden: bool = False, slug: Optional[str] = None):
-        try:
-            return {"files": wsr.tree_at(_read_target(request, slug), hidden=hidden)}
-        except ValueError:
-            raise HTTPException(status_code=400, detail="invalid subject")
 
-    @app.post("/api/workspace/upload")
-    async def ws_upload(request: Request, files: list[UploadFile] = File(...)):
-        if not files:
-            raise HTTPException(status_code=400, detail="no files uploaded")
-        subject = subject_of(request)
-        try:
-            ws = wsr.workspace_dir(subject)
-        except ValueError:
-            raise HTTPException(status_code=400, detail="invalid subject")
-        uploads = ws / "uploads"
-        uploads.mkdir(parents=True, exist_ok=True)
-        pending: list[tuple[Path, bytes, str, str]] = []
-        for file in files:
-            try:
-                content = await file.read()
-            finally:
-                await file.close()
-            if len(content) > MAX_UPLOAD_BYTES:
-                raise HTTPException(status_code=413, detail=f"{file.filename or 'upload'} exceeds 25MB")
-            safe_name = _upload_filename(file.filename)
-            digest = hashlib.sha256(content).hexdigest()
-            stored_name = f"{digest[:16]}-{safe_name}"
-            target = (uploads / stored_name).resolve()
-            if uploads.resolve() not in target.parents:
-                raise HTTPException(status_code=400, detail="invalid filename")
-            pending.append((target, content, stored_name, f"uploads/{stored_name}"))
-        uploaded: list[dict[str, str]] = []
-        for target, content, stored_name, path in pending:
-            target.write_bytes(content)
-            uploaded.append({"name": stored_name, "path": path})
-        return {"files": uploaded}
-
-    @app.get("/api/workspace/file")
-    def ws_file(request: Request, path: str, slug: Optional[str] = None):
-        try:
-            content = wsr.read_at(_read_target(request, slug), path)
-        except ValueError:
-            raise HTTPException(status_code=400, detail="invalid path")
-        if content is None:
-            raise HTTPException(status_code=404, detail="not found")
-        return {"path": path, "content": content}
-
-    @app.put("/api/workspace/file")
-    def ws_file_write(request: Request, body: dict = Body(...)):
-        """WRITE one doc — the terminal's in-place page editor (Codex-style). Authorization mirrors the
-        MOUNT rules, not the read rules: own baseline/_system always; a shared workspace needs
-        contributor+; `_global` only the admin allowlist. Commits so history stays honest."""
-        import shutil as _sh  # noqa: F401 — parity with ws_reset's import style
-        import subprocess as _sp
-        subject = subject_of(request)
-        rel = str(body.get("path") or "").strip()
-        slug = str(body.get("slug") or "").strip() or None
-        content = body.get("content")
-        if not isinstance(content, str):
-            raise HTTPException(status_code=400, detail="need a relative path and string content")
-        try:
-            wpaths.relative_parts(rel)   # absolute · `..` · `.git`/`.vexa` — one rule, every route
-        except wpaths.PathRefused as exc:
-            raise HTTPException(status_code=400, detail=str(exc)) from None
-        # `kg/templates/` is the SHAPE of an entity, not one. It is hidden from every tree, so a tab
-        # can only be open on it deliberately — and a save from that tab would silently rewrite the
-        # shape every future entity is copied from.
-        if rel.startswith("kg/templates/"):
-            raise HTTPException(status_code=403, detail="kg/templates/ holds entity shapes, not records")
-        if slug == system_mounts.GLOBAL_SLUG:
-            if not global_layer.is_admin(settings, str(subject)):
-                raise HTTPException(status_code=403, detail="only an org admin may edit _global")
-            candidates = [Path(settings.workspaces_dir) / system_mounts.GLOBAL_SLUG,
-                          Path(settings.global_system_workspace_path or "/nonexistent")]
-            target = next((c for c in candidates if c.is_dir() and os.access(c, os.W_OK)), None)
-            if target is None:
-                raise HTTPException(status_code=404, detail="the organisation tier is not writable here")
-        else:
-            if slug and slug not in (subject, system_mounts.SYSTEM_SLUG):
-                try:
-                    membership_mod.require_role(wsr.root, slug, subject, "contributor")
-                except MembershipError:
-                    pass  # not shared — fall through; _read_target 403s anything outside the active set
-            target = _read_target(request, slug, write=True)
-        try:
-            f = wpaths.resolve_inside(target, rel)   # …and again WITH the root, for the symlink half
-        except wpaths.PathRefused as exc:
-            raise HTTPException(status_code=400, detail=str(exc)) from None
-        f.parent.mkdir(parents=True, exist_ok=True)
-        f.write_text(content, encoding="utf-8")
-        if (target / ".git").is_dir():
-            _sp.run(["git", "-C", str(target), "add", rel], check=False, capture_output=True)
-            _sp.run(["git", "-C", str(target), "-c", "user.name=vexa-terminal", "-c", "user.email=terminal@vexa.local",
-                     "commit", "-m", f"edit {rel} (terminal page editor)"], check=False, capture_output=True)
-        return {"path": rel, "written": True}
 
     def _entity_mounts(subject: str) -> list:
         """`[{slug, path}]` for every workspace this subject has mounted — their own actives plus
@@ -2527,136 +772,13 @@ def create_app(
                 out.append({"slug": m.slug, "path": m.path})
         return out
 
-    @app.post("/api/workspace/entity")
-    def ws_entity_upsert(request: Request, body: dict = Body(...)):
-        """UPSERT one knowledge-graph entity — PRD decision 24, the single call behind `entity_upsert`.
 
-        Creates or updates `kg/entities/<kind>/<slug>.md` AS A CARD (decision 24.6): a summary, the
-        kind's sections, `## Connected` chips both ways, `## Sources`, `## Open questions`, and the
-        dated log at the end under `## Timeline`. A page in the old flat shape is re-rendered into
-        the card on its next touch, entries preserved. Then it refreshes `kg/INDEX.md` and commits
-        both by pathspec with the F31 subject shape. Authorization is `ws_file_write`'s, because it
-        is the same act — a write into a workspace — and two spellings of one authorization rule is
-        how the second one ends up weaker.
 
-        This endpoint exists so the MCP tool can be a THIN FORWARD (PRD §3.3: every host-reaching rig
-        tool is a missing endpoint in an owning service wearing a shell command). `workspace_write`'s
-        `docker exec` double is the shape this deliberately does not copy.
-        """
-        subject = subject_of(request)
-        kind = str(body.get("kind") or "").strip().lower()
-        name = str(body.get("name") or "").strip()
-        source = str(body.get("source") or "").strip()
-        slug = str(body.get("slug") or "").strip() or None
-        raw_facts = body.get("facts")
-        if isinstance(raw_facts, str):
-            raw_facts = [raw_facts]
-        facts = [str(f) for f in (raw_facts or []) if str(f).strip()]
-
-        if slug == system_mounts.GLOBAL_SLUG:
-            # `_global` is the company layer, admin-written and read-only to every worker. An entity
-            # write is exactly the kind of thing that would drift into it by accident.
-            raise HTTPException(status_code=403, detail="_global is the organisation tier — entities "
-                                                        "belong on a desk, not in it")
-        if slug and slug not in (subject, system_mounts.SYSTEM_SLUG):
-            try:
-                membership_mod.require_role(wsr.root, slug, subject, "contributor")
-            except MembershipError:
-                pass  # not shared — _read_target 403s anything outside the active set
-        target = _read_target(request, slug, write=True)
-        try:
-            # THE MOUNT SET IS HANDED IN so a `[[Name]]` whose page lives in another mounted
-            # workspace is stored as `[[ws:<id>/<entity-id>]]` (PRD decision 26.3). Without it the
-            # link resolves by TITLE in whichever mount the reader searches first, and dies the
-            # moment either workspace is renamed — which is the ordinary case, not the edge.
-            # `dates` (decision 31 §3): the whitelisted temporal frontmatter — `scheduled_at`,
-            # `held_at`, `report_delivered_at`. Filtered and normalised in `upsert_entity`, so a
-            # caller cannot write an arbitrary key by naming it here.
-            # THE CARD (decision 24.6). `summary`, `fields`, `section`, `connections` and
-            # `open_questions` are what turn an append into a page: a fact arrives knowing which
-            # section it belongs in, and a named company becomes an edge in both directions.
-            # Everything is optional — a caller that passes only `facts` still gets a card, with
-            # the facts in `## Timeline`, which is the shape the migration produces anyway.
-            result = entities_mod.upsert_entity(
-                target, kind, name, facts, source,
-                mounts=_entity_mounts(subject), dates=body.get("dates"),
-                summary=str(body.get("summary") or "").strip(),
-                fields=body.get("fields") if isinstance(body.get("fields"), dict) else None,
-                section=str(body.get("section") or "").strip(),
-                connections=body.get("connections") or (),
-                open_questions=body.get("open_questions") or ())
-        except entities_mod.EntityRefused as e:
-            # 422, not 400: the request is well-formed and the REFUSAL is the product — the agent is
-            # meant to read the sentence and fix the fact, not to retry the call.
-            raise HTTPException(status_code=422, detail=str(e))
-        index_rel = entities_mod.write_index(target, slug or str(subject))
-        sha = None
-        if result.get("changed"):
-            sha = entities_mod.commit_entity(
-                target, [result["path"], index_rel, *result.get("back_links", ())],
-                subject_path=result["path"], created=bool(result.get("created")),
-                author=(str(subject), f"{subject}@vexa.local"))
-        return {**result, "index": index_rel, "commit": sha}
-
-    @app.get("/api/workspace/git")
-    def ws_git(request: Request, slug: Optional[str] = None):
-        """Author-attributed source-control state (branch · working changes · recent commits) of a
-        workspace. No ``slug`` → the caller's own primary. A ``slug`` addresses a SHARED workspace the
-        caller is a member of (same authorized resolution as tree/file reads) — its commits carry
-        ``author`` + ``kind`` so the terminal can show OTHER members' agent pushes as they land."""
-        try:
-            target = _read_target(request, slug)  # authorizes: a slug outside the caller's mount set → 403
-            return wsr.git_state_at(target, viewer=subject_of(request))
-        except ValueError:
-            raise HTTPException(status_code=400, detail="invalid subject")
-
-    @app.get("/api/workspace/git/show")
-    def ws_git_show(request: Request, sha: str, slug: Optional[str] = None, path: Optional[str] = None):
-        """Unified diff of ONE commit (optionally one file) — same authorized resolution as ws_git — so
-        the terminal can highlight exactly what a commit changed. The optional ``path`` is a pathspec
-        and goes through the same guard as every other caller-supplied path."""
-        try:
-            target = _read_target(request, slug)  # authorizes: a slug outside the caller's mount set → 403
-            return wsr.git_diff_at(target, sha, path)
-        except ValueError:
-            raise HTTPException(status_code=400, detail="invalid path or subject")
 
     # ── workspace lifecycle (SCAFFOLD / TODO(phase-6)) — init from a validated template, swap which
     # validated workspace/template the next dispatch mounts. The seams exist downstream (seeding.seed_workspace
     # for init; VEXA_WORKSPACE_REPO/REF in dispatch/spawn for swap, bridge resolves per-meeting) — Phase 6
     # surfaces them here and wires the slim-client init_workspace()/use_workspace().
-    @app.post("/api/workspace/init", status_code=201)
-    def ws_init(request: Request):
-        """EAGERLY provision this subject's workspace tiers — the "on account creation" seam (so the
-        Personal baseline + the private `_system` tier exist BEFORE the first dispatch, instead of being
-        lazily seeded on first turn). Materializes the baseline from the VALIDATED workspace-seed template
-        (shared.seeding.seed_workspace) and ensures `_system` (system_mounts.ensure_system_workspace).
-        Idempotent — existing tiers (`.git` present) are returned untouched, so it's safe to call on every
-        login. The same seams the worker uses lazily on first dispatch, surfaced as a control."""
-        subject = subject_of(request)
-        ws = wsr.workspace_dir(subject)
-        # Select the seed out of the registry root (default template for now; per-request template
-        # selection lands with the second seed). VEXA_WORKSPACE_SEED_DIR still overrides.
-        seed_dir = resolve_seed_dir(
-            settings.default_template if settings is not None else None,
-            seeds_root=settings.workspace_seeds_dir if settings is not None else None,
-        )
-        problems = validate_seed(seed_dir)
-        if problems:
-            raise HTTPException(status_code=500, detail="invalid workspace seed: " + "; ".join(problems))
-        existed = (ws / ".git").exists()
-        seed_workspace(ws, seed_dir)
-        # The PRIVATE SYSTEM tier (`_system`) — always-mounted, holds the light identity reference. Ensure
-        # it up front too so identity + chats/settings have a home from the very first turn. Idempotent.
-        system_existed = (system_mounts.system_store_path(wsr.root, subject) / ".git").exists()
-        system_mounts.ensure_system_workspace(str(wsr.root), subject)
-        # The desk's id + registry row. Named by the address that signed in when the gateway gave
-        # us one: a desk called "Desk 126" is better than `126` and worse than the person's own
-        # address, and this is the one seam that has the address.
-        _ws_sync(str(subject), kind="desk", owner=str(subject), ws_dir=ws,
-                 name=(request.headers.get("x-user-email") or "").strip() or None)
-        return {"workspace": str(ws), "seeded": not existed, "already_initialized": existed,
-                "system_seeded": not system_existed}
 
     # ── workspace identity + link resolution (PRD decision 26) ──────────────────────────────────
     #
@@ -2695,517 +817,27 @@ def create_app(
         access = ids_mod.access_for(rec, subject, root=wsr.root, is_member=_ws_is_member)
         return rec if access == ids_mod.ACCESS_READABLE else None
 
-    @app.get("/api/workspaces/by-slug/{slug}")
-    def ws_id_by_slug(slug: str, request: Request):
-        """The identity of a workspace addressed the OLD way — by slug. What the terminal calls to
-        put a NAME where it used to print a directory name (F49: the chat header read `126`)."""
-        subject = subject_of(request)
-        rec = workspace_registry.by_slug(slug) or _ws_sync(slug)
-        access = ids_mod.access_for(rec, subject, root=wsr.root, is_member=_ws_is_member)
-        return ids_mod.view(rec, access, writable=ids_mod.writable_for(
-            rec, subject, root=wsr.root, is_member=_ws_is_member))
 
-    @app.get("/api/workspaces/{workspace_id}")
-    def ws_id_resolve(workspace_id: str, request: Request):
-        """`{id, name, kind, access}` for one workspace id, from THIS reader's point of view.
 
-        Never 404s and never 403s: `not-yours` and `gone` are ANSWERS (decision 26.3), and a status
-        code would make the client render an error where the design says render a greyed chip."""
-        subject = subject_of(request)
-        rec = workspace_registry.get(workspace_id)
-        access = ids_mod.access_for(rec, subject, root=wsr.root, is_member=_ws_is_member)
-        return ids_mod.view(rec, access, workspace_id=workspace_id, writable=ids_mod.writable_for(
-            rec, subject, root=wsr.root, is_member=_ws_is_member))
 
-    @app.post("/api/workspaces/{workspace_id}/rename")
-    def ws_id_rename(workspace_id: str, request: Request, body: dict = Body(default={})):
-        """Rename a workspace. The id does not move, and neither does anything pointing at it —
-        that is the single behaviour PRD decision 26 was asked for, and this is the route that
-        exercises it.
 
-        WHO (founder ruling, 2026-09-02): the group's OWNER, and instance admins. Nobody else, and
-        deliberately not a desk's own owner — a desk's name comes from the address that signed in,
-        and renaming one is a question about identity display that has not been asked yet.
 
-        AUDITED: who, old, new, when, kept on the record (capped) and logged. A rename is the one
-        operation whose whole point is that nothing else changes, which means the only way to see
-        that it happened at all is to have written it down."""
-        subject = subject_of(request)
-        try:
-            rec = ids_mod.rename_audited(
-                workspace_registry, workspace_id, str(body.get("name") or ""), by=str(subject),
-                is_admin=bool(global_layer.is_admin(settings, str(subject))),
-                root=wsr.root, is_member=_ws_is_member)
-        except KeyError:
-            raise HTTPException(status_code=404, detail="no workspace with that id")
-        except ids_mod.RenameRefused as e:
-            raise HTTPException(status_code=403, detail=str(e))
-        except ValueError as e:
-            raise HTTPException(status_code=400, detail=str(e))
-        return {**ids_mod.view(rec, ids_mod.ACCESS_READABLE, writable=ids_mod.writable_for(
-            rec, subject, root=wsr.root, is_member=_ws_is_member)),
-            "renamed_from": (rec.get("renames") or [{}])[-1].get("from")}
 
-    @app.post("/api/desk/touch", status_code=202)
-    def desk_touch(request: Request, body: dict = Body(default={})):
-        """Record that this person opened one page — the ranking signal behind the desk README.
-
-        202, not 200: it is a report, the caller has nothing to do with the answer, and a panel
-        must never wait on bookkeeping to render a document.
-
-        WHAT IT WILL NOT DO is trust the workspace it is told. The body names a workspace id; the
-        access rule is applied to it exactly as it is for a link, so a touch cannot be used to ask
-        whether a workspace exists, and a page in a workspace this caller cannot read is not
-        recorded. The DESK the touch is filed under is always the CALLER'S OWN — never a parameter,
-        because "whose desk does this belong to" is not a question a client gets to answer."""
-        subject = subject_of(request)
-        wid = str(body.get("workspace") or "").strip()
-        # NOT `.lstrip("/")`. Silently turning `/etc/passwd` into `etc/passwd` is a normalization
-        # that RESCUES the one input the guard below exists to refuse — and this route's answer
-        # travels into the desk README as a link.
-        path = str(body.get("path") or "").strip()
-        try:
-            wpaths.relative_parts(path)
-        except wpaths.PathRefused as exc:
-            raise HTTPException(status_code=400, detail=str(exc)) from None
-        if not wid:
-            raise HTTPException(status_code=400, detail="a touch names a workspace id and a path")
-        rec = workspace_registry.get(wid)
-        if ids_mod.access_for(rec, subject, root=wsr.root, is_member=_ws_is_member) != ids_mod.ACCESS_READABLE:
-            return {"recorded": False}          # not readable — nothing to say, and nothing leaked
-        desk = workspace_registry.by_slug(str(subject)) or _ws_sync(str(subject), kind="desk",
-                                                                    owner=str(subject))
-        if not desk:
-            return {"recorded": False}
-        workspace_touches.touch(desk["id"], wid, path)
-        # Mirror it where the WORKER can read it: the README is regenerated at the end of a turn,
-        # in a container that holds the mounts and no redis.
-        ids_mod.mirror_touches(desk.get("dir") or (wsr.root / str(subject)),
-                               workspace_touches.recent(desk["id"]))
-        return {"recorded": True}
-
-    @app.post("/api/links/resolve")
-    def links_resolve(request: Request, body: dict = Body(default={})):
-        """A page's refs → what each one points at now, per ref: `{ref, title, url, access}`.
-
-        One round trip for a whole document on purpose. The panel renders a doc at once, and a
-        request per link would be a burst of them against the same three directories."""
-        refs = body.get("refs")
-        if not isinstance(refs, list):
-            raise HTTPException(status_code=400, detail="refs must be a list of link references")
-        slug = str(body.get("slug") or "").strip() or None
-        return {"results": link_resolver_mod.resolve_many(
-            [str(r) for r in refs], subject=subject_of(request), root=wsr.root,
-            registry=workspace_registry, here=_ws_here(request, slug), is_member=_ws_is_member)}
-
-    @app.get("/api/workspace/attached")
-    def ws_attached(request: Request):
-        """The subject's attachment view: the active slug + the parked workspaces available to swap back
-        to, plus ``published_url`` — where the ACTIVE workspace was published (the ``vexa-publish``
-        remote's token-free URL), or null when it never was. The client renders a published workspace
-        with a link to its GitHub home instead of the publish action."""
-        subject = subject_of(request)
-        try:
-            state = attached_workspaces(wsr.root, subject)
-        except ValueError:
-            raise HTTPException(status_code=400, detail="invalid subject")
-        state["published_url"] = published_remote_url(wsr.workspace_dir(subject))
-        return state
-
-    @app.post("/api/workspace/swap")
-    def ws_swap(request: Request, body: WorkspaceSwapBody = Body(default=WorkspaceSwapBody())):
-        """Attach a CUSTOM external git repo as this subject's active workspace (swap). The currently
-        active workspace is PARKED (kept, never destroyed) so it can be swapped back to; the requested
-        repo is restored from a prior park or cloned fresh. Omit ``repo`` to swap back to the seed.
-
-        Mounting is by-folder (``<root>/<subject>`` is what the next dispatch mounts), so the swapped
-        tree takes effect on the subject's next turn — no dispatch change needed."""
-        subject = subject_of(request)
-        repo = _repo(body.repo)      # 422 before any git process exists
-        key = deploy_keys_mod.workspace_key(subject=subject)
-        try:
-            with wcreds.for_workspace(wsr.root, key=key, repo_url=repo or "", subject=subject,
-                                      explicit_token=body.token) as cred:
-                result = swap_workspace(wsr.root, subject, repo, body.ref or "main",
-                                        slug=body.slug or None, fresh=body.fresh, token=cred.token,
-                                        clone=_clone_fn(cred))
-        except repo_ref.RepoRefError as exc:
-            raise HTTPException(status_code=422, detail=exc.sentence)
-        except ValueError:
-            raise HTTPException(status_code=400, detail="invalid subject")
-        except KeyError:
-            raise HTTPException(status_code=404, detail="unknown workspace")
-        except CloneError as exc:
-            # Already token-redacted (P15). A private repo we hold no credential for lands here — and
-            # the answer is the workspace's public key to add, not a prompt for a secret.
-            raise _credential_refusal(f"git clone failed: {redact_secrets(exc)}", subject, None, repo or "")
-        return {
-            "subject": result.subject,
-            "active": result.active_slug,
-            "repo": result.repo,
-            "ref": result.ref,
-            "swapped": result.swapped,
-            "cloned": result.cloned,
-            "parked": result.parked_slug,
-            "nested": result.nested,
-        }
 
     # ── the additive mount set (WP-A2.1): ACTIVE-SET membership over swap's park/restore machinery ──────
-    @app.get("/api/workspace/active")
-    def ws_active(request: Request):
-        """The subject's ordered ACTIVE SET — the workspaces the next dispatch mounts (the private baseline
-        first, then any activated extras). Each: ``slug, repo, ref, role, path, write, primary``."""
-        subject = subject_of(request)
-        try:
-            mounts = active_workspaces(wsr.root, subject)
-        except ValueError:
-            raise HTTPException(status_code=400, detail="invalid subject")
-        # Lane A: append the SHARED workspaces the subject is a member of. Enumeration reconciles BOTH
-        # membership stores (index ∪ policy/members.json — the same union as the shared listing) so a dead
-        # or incomplete index cannot silently drop a locally-held grant from the mount set;
-        # shared_active_mounts still re-checks the role authoritatively per workspace. A resolution failure
-        # costs the "shared" section of the set, never the subject's own private mounts — and
-        # ``index_degraded`` says out loud when the mirror could not be read.
-        rows, index_degraded = membership_mod.reconciled_memberships(wsr.root, subject, mindex.list)
-        try:
-            mounts = mounts + shared_active_mounts(wsr.root, subject, rows)
-        except Exception:  # noqa: BLE001 — a shared-mount resolution hiccup must not break the active-set read
-            logger.warning("shared-mount resolution failed for subject=%s — returning private mounts only", subject)
-        return {
-            "subject": subject,
-            "active": [
-                {"slug": m.slug, "repo": m.repo, "ref": m.ref, "role": m.role,
-                 "path": m.path, "write": m.write, "primary": m.primary, "name": m.name}
-                for m in mounts
-            ],
-            "index_degraded": index_degraded,
-        }
 
-    @app.post("/api/workspace/activate")
-    def ws_activate(request: Request, body: WorkspaceActivateBody = Body(default=WorkspaceActivateBody())):
-        """ADD a workspace to the active set WITHOUT parking the others (the additive counterpart of swap).
-        Clones/restores the target if needed. Idempotent — an already-active workspace is a no-op."""
-        subject = subject_of(request)
-        repo = _repo(body.repo)      # 422 before any git process exists
-        key = deploy_keys_mod.workspace_key(subject=subject)
-        try:
-            with wcreds.for_workspace(wsr.root, key=key, repo_url=repo or "", subject=subject,
-                                      explicit_token=body.token) as cred:
-                result = activate_workspace(wsr.root, subject, repo, body.ref or "main",
-                                            slug=body.slug or None, token=cred.token,
-                                            clone=_clone_fn(cred))
-        except repo_ref.RepoRefError as exc:
-            raise HTTPException(status_code=422, detail=exc.sentence)
-        except ValueError:
-            raise HTTPException(status_code=400, detail="invalid subject")
-        except KeyError:
-            raise HTTPException(status_code=404, detail="unknown workspace")
-        except CloneError as exc:
-            raise _credential_refusal(f"git clone failed: {redact_secrets(exc)}", subject, None, repo or "")
-        return {"subject": result.subject, "slug": result.slug, "changed": result.changed,
-                "cloned": result.cloned, "nested": result.nested}
 
-    @app.post("/api/workspace/new", status_code=201)
-    def ws_new(request: Request, body: WorkspaceNewBody = Body(default=WorkspaceNewBody())):
-        """CREATE a brand-new BLANK workspace (seeded from the template) at a fresh unique slug and ADD it
-        to the active set (additive — the "new workspace" action). Nothing is parked/rebuilt/backed up: the
-        private baseline and every other active workspace stay exactly as they were."""
-        subject = subject_of(request)
-        try:
-            result = create_workspace(wsr.root, subject, name=body.name or None)
-        except ValueError:
-            raise HTTPException(status_code=400, detail="invalid subject")
-        return {"subject": result.subject, "slug": result.slug, "changed": result.changed,
-                "added": True}
 
-    @app.post("/api/workspace/deactivate")
-    def ws_deactivate(request: Request, body: WorkspaceDeactivateBody = Body(...)):
-        """REMOVE a workspace from the active set (park it — never destroyed). The private baseline can be
-        switched off too (sets ``baseline_hidden``; its home tree is untouched, re-activate to switch it back
-        on). Idempotent — an already-off / not-active slug is a no-op."""
-        subject = subject_of(request)
-        try:
-            result = deactivate_workspace(wsr.root, subject, body.slug)
-        except ValueError:
-            raise HTTPException(status_code=400, detail="invalid subject")
-        return {"subject": result.subject, "slug": result.slug, "changed": result.changed}
 
-    @app.post("/api/workspace/publish")
-    def ws_publish(request: Request, body: WorkspacePublishBody = Body(...)):
-        """Publish this subject's vexa-born workspace to GitHub — the counterpart of swap/attach.
-        Creates the repo under the caller's account (or ``org``) with their per-call PAT, then pushes
-        the active workspace's current branch (FULL history) over the token-scrubbed dedicated remote.
-        ``remote_url`` skips creation (pre-created/empty repo). Re-publish = plain push (fast-forward
-        or a clear error on divergence — never a force push). The token is used server-side for this
-        call only and never stored; every error is token-redacted (P15)."""
-        subject = subject_of(request)
-        token = (body.token or "").strip() or git_creds.read_github_token(wsr.root, subject)
-        if not token:
-            raise HTTPException(status_code=400, detail="a GitHub token is required — pass one or save a reusable token")
-        try:
-            result = publish_workspace(
-                wsr.root, subject,
-                token=token, repo_name=body.repo_name, private=body.private,
-                org=body.org or None, remote_url=body.remote_url or None,
-                # slug → any workspace the caller can manage (own parked slot or shared membership,
-                # resolved + permission-checked by _manage_dir); omitted keeps the legacy seed target.
-                ws_dir=_manage_dir(subject, body.slug) if body.slug else None,
-            )
-        except ValueError as exc:
-            raise HTTPException(status_code=400, detail=str(exc) or "invalid subject")
-        except RepoExistsError as exc:
-            raise HTTPException(status_code=409, detail=str(exc))   # already token-redacted (P15)
-        except PublishError as exc:
-            raise HTTPException(status_code=502, detail=str(exc))   # already token-redacted (P15)
-        return {
-            "repo_url": result.repo_url,
-            "pushed_ref": result.pushed_ref,
-            "head_sha": result.head_sha,
-            "created": result.created,
-        }
 
-    @app.post("/api/workspace/rename")
-    def ws_rename(request: Request, body: WorkspaceRenameBody = Body(...)):
-        """Rename a workspace slot — a DISPLAY label only. The slug and the parked tree are unchanged, so
-        swap-back and repo re-attach keep matching. Pass an empty ``name`` to clear the label."""
-        subject = subject_of(request)
-        try:
-            return rename_workspace(wsr.root, subject, body.slug, body.name)
-        except ValueError:
-            raise HTTPException(status_code=400, detail="invalid subject")
-        except KeyError:
-            raise HTTPException(status_code=404, detail="unknown workspace")
 
-    @app.get("/api/workspace/git-token")
-    def ws_git_token_get(request: Request):
-        """Whether the caller has a SAVED reusable GitHub token, and a masked (last-4) preview of it. The
-        clear value is NEVER returned — server-side only (git_credentials)."""
-        subject = subject_of(request)
-        return {"set": git_creds.read_github_token(wsr.root, subject) is not None,
-                "masked": git_creds.masked_github_token(wsr.root, subject)}
 
-    @app.post("/api/workspace/git-token")
-    def ws_git_token_set(request: Request, body: GitTokenBody = Body(default=GitTokenBody())):
-        """Save (or CLEAR, with an empty token) the caller's reusable GitHub token — stored once, server-
-        side, and applied as the fallback credential for every git op across all their repos. Returns the
-        masked state, never the clear value."""
-        subject = subject_of(request)
-        try:
-            stored = git_creds.set_github_token(wsr.root, subject, body.token)
-        except ValueError as exc:
-            raise HTTPException(status_code=400, detail=str(exc))
-        return {"set": stored, "masked": git_creds.masked_github_token(wsr.root, subject)}
 
-    @app.get("/api/workspace/git-remote-status")
-    def ws_git_remote_status(request: Request, slug: Optional[str] = None):
-        """The GitHub-sync state of a workspace (default = the caller's primary; ``slug`` = one of their
-        own or shared workspaces). Read-only + no network: reports the home remote (origin / vexa-publish),
-        its URL, the branch, and ahead/behind counts vs the last-fetched tracking ref. No token needed."""
-        subject = subject_of(request)
-        ws = _manage_dir(subject, slug)
-        s = remote_status(ws)
-        return {
-            "has_home": s.has_home, "remote": s.remote, "url": s.url, "branch": s.branch,
-            "tracked": s.tracked, "ahead": s.ahead, "behind": s.behind,
-        }
 
-    @app.post("/api/workspace/push")
-    def ws_push(request: Request, body: WorkspacePushBody = Body(...)):
-        """Push a workspace's current branch to its GitHub home (origin for attached clones, vexa-publish
-        for published vexa-born), fast-forward only — NEVER a force push. The token authenticates the push
-        and is never stored; a diverged remote fails loud (pull first). Every error is token-redacted (P15)."""
-        subject = subject_of(request)
-        ws = _manage_dir(subject, body.slug)
-        home = remote_status(ws)
-        key = _workspace_key(subject, body.slug)
-        try:
-            with wcreds.for_workspace(wsr.root, key=key, repo_url=home.url or "", subject=subject,
-                                      explicit_token=body.token) as cred:
-                if cred.kind == "none":
-                    raise HTTPException(status_code=400, detail=(
-                        "this workspace has no credential — add its deploy key to the repository "
-                        "(POST /api/workspace/{slug}/deploy-key) or save a reusable GitHub token"))
-                r = push_origin(ws, token=cred.token, ssh_env=cred.ssh_env)
-        except ValueError as exc:
-            raise HTTPException(status_code=400, detail=str(exc))
-        except RemoteSyncError as exc:
-            raise _credential_refusal(redact_secrets(exc), subject, body.slug, home.url or "")  # P15
-        return {"remote": r.remote, "url": r.url, "branch": r.branch, "head_sha": r.head_sha}
 
-    @app.post("/api/workspace/pull")
-    def ws_pull(request: Request, body: WorkspacePullBody = Body(default=WorkspacePullBody())):
-        """Fetch + FAST-FORWARD a workspace from its GitHub home. A divergence (local commits the remote
-        lacks) is refused — no merge/rebase/force — so it is resolved deliberately. The token (optional for
-        public repos) is used for the fetch only and never stored (P15)."""
-        subject = subject_of(request)
-        ws = _manage_dir(subject, body.slug)
-        # A pull REWRITES the tree, so on a shared workspace it is a write: viewers are refused here even
-        # though they may read the same workspace through _manage_dir.
-        _require_shared_write(subject, body.slug)
-        home = remote_status(ws)
-        key = _workspace_key(subject, body.slug)
-        try:
-            with wcreds.for_workspace(wsr.root, key=key, repo_url=home.url or "", subject=subject,
-                                      explicit_token=body.token) as cred:
-                r = pull_origin(ws, token=cred.token, ssh_env=cred.ssh_env)  # None ⇒ public-repo fetch
-        except ValueError as exc:
-            raise HTTPException(status_code=400, detail=str(exc))
-        except RemoteSyncError as exc:
-            raise _credential_refusal(redact_secrets(exc), subject, body.slug, home.url or "")  # P15
-        return {"remote": r.remote, "url": r.url, "branch": r.branch, "head_sha": r.head_sha,
-                "updated": r.updated, "behind_before": r.behind_before}
 
-    @app.get("/api/workspace/purpose")
-    def ws_purpose_get(request: Request, slug: Optional[str] = None):
-        """Read a workspace's PURPOSE one-liner (default = the caller's primary; ``slug`` = one of their
-        own or shared workspaces). ``""`` when unset."""
-        subject = subject_of(request)
-        ws = _manage_dir(subject, slug)
-        return {"purpose": read_purpose(ws)}
 
-    @app.post("/api/workspace/purpose")
-    def ws_purpose_set(request: Request, body: WorkspacePurposeBody = Body(default=WorkspacePurposeBody())):
-        """Set (or clear) a workspace's PURPOSE — stored in the workspace + committed so it travels when
-        shared and feeds the mount preamble. Returns the normalized purpose actually stored."""
-        subject = subject_of(request)
-        ws = _manage_dir(subject, body.slug)
-        return {"purpose": write_purpose(ws, body.purpose)}
 
-    @app.get("/api/meeting/stream")
-    def meeting_stream(meeting_id: str, session_uid: str, request: Request):
-        """SSE feed for a LIVE meeting: the transcript Stream (`tc:meeting:{id}`), and only that.
-        Segments as the bot hears them, `retract` when the collector withdraws a draft, and
-        `meeting-end` when the session closes. It used to MERGE a second stream — the copilot's
-        output (`unit:agent-meet-{sid}:out`: `card`s, `note`s, `model-error`, `message-delta`) —
-        and PRD decision 34 removed the producer of all of it, so the merge went with it.
-
-        RESUMABLE: every event carries an SSE ``id:`` = the per-stream redis cursors. On reconnect the
-        browser echoes the last one as ``Last-Event-ID``; we resume EXACTLY from there (redis streams are
-        durable + id-addressable) instead of re-seeding only the last N entries. Without this, a transient
-        disconnect (the 'Live stream disconnected — reconnecting' path) dropped every segment published in
-        the gap beyond the bounded replay window from the LIVE view — the real-time transcript-loss bug
-        (the durable store kept them, so they only reappeared post-time)."""
-        if not redis_url:
-            raise HTTPException(status_code=501, detail="redis not wired")
-
-        # P0 (cross-tenant leak fix — SSE sibling of the by-id REST ownership check): OWNER-SCOPE the live
-        # feed BEFORE opening any redis stream. `meeting_id` (row id) + `session_uid` arrive from the
-        # caller's query params; row ids are sequential ints, so without this an authenticated user B could
-        # `EventSource(...?meeting_id=<A_row>&session_uid=<A_native>)` and stream tenant A's live transcript
-        # (an ACTIVE, enumerable cross-tenant read). Mirror the WS `/ws` path: derive the
-        # caller identity (`subject_of` → 401 on no gateway-injected X-User-Id) and verify the caller OWNS
-        # the requested row (meeting-api `GET /meetings/{id}` owner-scopes in SQL: `Meeting.user_id ==
-        # user_id` → 404 for a foreign/absent row). Fail CLOSED (403) BEFORE the stream opens.
-        # OWNER-ONLY for now (matches the WS path today); a shared-workspace membership grant would extend
-        # `_meeting_owner_lookup` — the clean seam — but is intentionally NOT honored here yet.
-        subject = subject_of(request)  # 401 if no (gateway-injected) identity — fail closed
-        owned = _meeting_owner_lookup(subject, meeting_id)
-        if owned is None:
-            # Absent row, or a row owned by a DIFFERENT tenant → refuse (404-equivalent, no stream opened).
-            raise HTTPException(status_code=403, detail="not authorized for this meeting")
-        # `session_uid` is ALSO caller-supplied. The terminal passes the ROW id as `session_uid` for
-        # live rows (liveMeetings.ts `session_uid = live ? id : undefined`); the meeting's own native
-        # id is accepted for the legacy shape (native==row==session). Bind it to the OWNED row so a
-        # caller cannot pair its own row with someone else's key.
-        owned_native = str(owned.get("native_meeting_id") or "")
-        if session_uid not in (owned_native, str(meeting_id)):
-            raise HTTPException(status_code=403, detail="session_uid does not match this meeting")
-
-        resume_t = _decode_sse_cursor(request.headers.get("last-event-id"))
-
-        def gen():
-            import redis
-
-            r = redis.from_url(redis_url, decode_responses=True)
-            tkey = f"tc:meeting:{meeting_id}"
-            # Resume EXACTLY from the client's last-seen cursor when present (gapless reconnect);
-            # otherwise seed then live-tail (fresh connect).
-            last = {tkey: resume_t or "$"}
-            idle = 0
-            # A meeting row's transcript stream is REUSED across the meeting's sessions, so a
-            # `session_end` is not necessarily the end of the VIEW: a new session can resume on the
-            # same key moments later, and closing on the first marker painted "Meeting ended" over a
-            # live meeting. So an end ARMS a short drain (one poll) and a `session_start` or any real
-            # segment disarms it. This used to be a 45s bounded drain, because the copilot's final
-            # beat ran ~10s after session_end and its notes had to reach the view first; with that
-            # producer gone (PRD decision 34) the only thing that can still arrive is a resumed
-            # session, and one poll is enough to see it.
-            ending = False
-
-            def cursor():
-                return _encode_sse_cursor(last, tkey)
-
-            def seg_events(payload):
-                for seg in payload.get("segments", []):
-                    yield ({"type": "transcript", "speaker": seg.get("speaker"),
-                            "text": seg.get("text"), "t": seg.get("start"),
-                            "tsMs": seg.get("abs_start_ms"),
-                            "completed": seg.get("completed", True),
-                            "id": seg.get("segment_id")}, cursor())
-
-            def retract_event(payload):
-                """A `retract` marker (the collector withdrew superseded/over-extended pending drafts) →
-                a `retract` SSE event so the terminal drops those segment ids from the live view."""
-                ids = payload.get("segment_ids") or []
-                if ids:
-                    yield ({"type": "retract", "segment_ids": ids}, cursor())
-
-            if resume_t is None:   # fresh connect → seed the bounded recent transcript tail
-                seed_rows = list(reversed(r.xrevrange(tkey, count=MEETING_STREAM_TRANSCRIPT_REPLAY) or []))
-                for entry_id, fields in seed_rows:
-                    last[tkey] = entry_id
-                    payload = json.loads(fields.get("payload", "{}"))
-                    if payload.get("type") == "session_end":
-                        ending = True
-                        last.pop(tkey, None)
-                        continue
-                    if payload.get("type") == "retract":
-                        yield from retract_event(payload)
-                        continue
-                    # A real segment AFTER a session_end in the replay tail means a NEW session resumed
-                    # on this reused meeting-row stream (tc:meeting:{id} is shared across a meeting's
-                    # sessions). The prior end must NOT close the current live view.
-                    ending = False
-                    yield from seg_events(payload)
-
-            while True:
-                resp = r.xread(last, count=500, block=1500 if ending else 15000)
-                if not resp:
-                    if ending:
-                        live.drop(session_uid)  # leaves the terminal's live-meetings feed
-                        yield ({"type": "meeting-end"}, cursor())
-                        return
-                    idle += 15000
-                    if idle >= 600000:
-                        return
-                    yield ({"type": "ping"}, cursor())
-                    continue
-                idle = 0
-                for stream, entries in resp:
-                    for entry_id, fields in entries:
-                        last[stream] = entry_id
-                        if stream == tkey:
-                            payload = json.loads(fields.get("payload", "{}"))
-                            ptype = payload.get("type")
-                            if ptype == "session_end":
-                                ending = True            # a resumed session gets one poll to appear
-                                last.pop(tkey, None)     # session_end is the last transcript entry
-                                break
-                            if ptype == "retract":
-                                yield from retract_event(payload)
-                                continue
-                            # A `session_start` marker — or any real segment — after a prior
-                            # session_end means the meeting is LIVE again on this reused row: clear
-                            # the stale arm so the drain never fires a premature meeting-end.
-                            ending = False
-                            if ptype == "session_start":
-                                continue                 # marker only — nothing to render
-                            yield from seg_events(payload)
-
-        return StreamingResponse(
-            _sse(gen()), media_type="text/event-stream",
-            headers={"Cache-Control": "no-cache", "X-Accel-Buffering": "no"},
-        )
 
 
     # ── workspace membership + invites + roles (Lane M) ───────────────────────────────────────────
@@ -3246,164 +878,13 @@ def create_app(
             return HTTPException(status_code=502, detail=detail)
         return HTTPException(status_code=502, detail=f"{detail}\n\n{wcreds.prompt_sentence(prompt)}")
 
-    @app.post("/api/workspace/shared/{workspace_id}/attach")
-    def ws_shared_attach(workspace_id: str, request: Request, body: SharedAttachBody = Body(default=SharedAttachBody())):
-        """Attach an EXISTING git repo as a shared workspace's tree. Contributor or owner only — a
-        viewer may read the group's workspace, never replace it.
 
-        Park-and-clone, exactly as the desk swap: the current tree is kept under the workspace's own
-        store and can be swapped back to by slug with no re-clone, so this is reversible. ``policy/``
-        (the member list) is carried into the new tree, so an attach can never lock the group out."""
-        subject = subject_of(request)
-        try:
-            membership_mod.require_role(wsr.root, workspace_id, subject, "contributor")
-        except MembershipError as exc:
-            raise _member_error(exc)
-        repo = _repo(body.repo)      # 422 before any git process exists
-        key = deploy_keys_mod.workspace_key(workspace_id=workspace_id)
-        try:
-            with wcreds.for_workspace(wsr.root, key=key, repo_url=repo or "", subject=subject,
-                                      explicit_token=body.token) as cred:
-                clone = _clone_fn(cred)
-                result = attach_shared_workspace(wsr.root, workspace_id, repo, body.ref or "main",
-                                                 slug=body.slug or None, token=cred.token, clone=clone)
-        except repo_ref.RepoRefError as exc:
-            raise HTTPException(status_code=422, detail=exc.sentence)
-        except ValueError:
-            raise HTTPException(status_code=400, detail="invalid workspace id")
-        except KeyError:
-            raise HTTPException(status_code=404, detail="unknown workspace")
-        except CloneError as exc:   # message already token-redacted (P15)
-            raise _credential_refusal(f"git clone failed: {redact_secrets(exc)}", subject, workspace_id, repo or "")
-        return {
-            "workspace_id": workspace_id, "active": result.active_slug, "repo": result.repo,
-            "ref": result.ref, "attached": result.swapped, "cloned": result.cloned,
-            "parked": result.parked_slug, "nested": result.nested,
-            "state": ("cloned" if result.cloned else "restored" if result.swapped else "already attached"),
-        }
 
-    @app.get("/api/workspace/shared/{workspace_id}/attached")
-    def ws_shared_attached(workspace_id: str, request: Request):
-        """A shared workspace's attachment view — the active slug and the parked trees available to swap
-        back to, plus its GitHub home. Any member may read it (it says WHAT is mounted, never a credential)."""
-        subject = subject_of(request)
-        try:
-            membership_mod.require_role(wsr.root, workspace_id, subject, "viewer")
-        except MembershipError as exc:
-            raise _member_error(exc)
-        try:
-            state = shared_attached_state(wsr.root, workspace_id)
-        except ValueError:
-            raise HTTPException(status_code=400, detail="invalid workspace id")
-        ws = membership_mod._ws_dir(wsr.root, workspace_id)
-        st = remote_status(ws)
-        key = deploy_keys_mod.workspace_key(workspace_id=workspace_id)
-        state["home"] = {"remote": st.remote, "url": st.url, "branch": st.branch,
-                         "ahead": st.ahead, "behind": st.behind}
-        # A CAPABILITY, not a secret: whether a credential exists and of what kind — never its value.
-        state["credential"] = wcreds.home_capability(wsr.root, key=key, remote=st.remote, url=st.url,
-                                                    subject=subject)
-        return state
 
-    @app.get("/api/workspace/{slug}/deploy-key")
-    def ws_deploy_key_get(slug: str, request: Request):
-        """This workspace's PUBLIC deploy key (null when none has been generated). The private half has
-        no read path at all — it is sealed in the credential store and materialized only for one git op."""
-        subject = subject_of(request)
-        _manage_dir(subject, slug)          # authorization: own slot, or a workspace they belong to
-        key = _workspace_key(subject, slug)
-        pub = deploy_keys_mod.public_key(wsr.root, key)
-        return {"slug": slug, "public_key": pub, "fingerprint": deploy_keys_mod.fingerprint(pub),
-                "add_as": "a deploy key with WRITE access"}
 
-    @app.post("/api/workspace/{slug}/deploy-key")
-    def ws_deploy_key_ensure(slug: str, request: Request, body: dict = Body(default={})):
-        """Generate this workspace's deploy key (idempotent — a second call returns the SAME key, so a
-        key the person already added to their repo is never invalidated) and say where to add it.
 
-        This is the credential model: they add our PUBLIC key to their repository; nothing of theirs
-        ever travels to us, and the private half is sealed at rest and never leaves this server."""
-        subject = subject_of(request)
-        _manage_dir(subject, slug)
-        repo_url = str((body or {}).get("repo") or "")
-        try:
-            prompt = wcreds.deploy_key_prompt(wsr.root, key=_workspace_key(subject, slug), repo_url=repo_url)
-        except deploy_keys_mod.DeployKeyError as exc:
-            raise HTTPException(status_code=501, detail=str(exc))
-        except ValueError:
-            raise HTTPException(status_code=400, detail="invalid workspace")
-        return {"slug": slug, **prompt, "message": wcreds.prompt_sentence(prompt)}
 
-    @app.post("/api/workspace/shared/{workspace_id}/active")
-    def ws_shared_active(workspace_id: str, request: Request, body: SharedActiveBody = Body(...)):
-        """Switch a SHARED workspace ON/OFF in the caller's active set (mount vs hide). Membership is
-        unchanged — this is a per-user mount preference so a member can 'switch it off' without leaving."""
-        subject = subject_of(request)
-        try:
-            set_shared_active(wsr.root, subject, workspace_id, body.active)
-        except ValueError:
-            raise HTTPException(status_code=400, detail="invalid workspace")
-        return {"workspace_id": workspace_id, "active": body.active}
 
-    @app.post("/api/workspace/{slug}/archive")
-    def ws_archive(slug: str, request: Request, body: ArchiveBody = Body(default=ArchiveBody())):
-        """Archive (collapse, keep the data) or un-archive one of the caller's own workspaces."""
-        subject = subject_of(request)
-        try:
-            set_archived(wsr.root, subject, slug, body.archived)
-        except ValueError as exc:
-            raise HTTPException(status_code=400, detail=str(exc))
-        except KeyError:
-            raise HTTPException(status_code=404, detail="workspace not found")
-        return {"slug": slug, "archived": body.archived}
-
-    @app.delete("/api/workspace/{slug}")
-    def ws_delete(slug: str, request: Request):
-        """DELETE one of the caller's own workspaces — removes the data irreversibly. Baseline is refused."""
-        subject = subject_of(request)
-        try:
-            delete_workspace(wsr.root, subject, slug)
-        except ValueError as exc:
-            raise HTTPException(status_code=400, detail=str(exc))
-        except KeyError:
-            raise HTTPException(status_code=404, detail="workspace not found")
-        return {"slug": slug, "deleted": True}
-
-    @app.post("/api/workspace/reset")
-    def ws_reset(request: Request, body: dict = Body(...)):
-        """RESET a structural folder to the seed — the delete-equivalent for tiers whose SLOT must
-        survive: the caller's `personal` baseline, or `_global` (admin allowlist only). Both are
-        "just folders" (founder ruling 2026-08-22) — wipe the content, re-copy the seed, commit.
-        `_system` is deliberately NOT resettable: it is sessions/continuity, not knowledge."""
-        import shutil as _sh
-        import subprocess as _sp
-        subject = subject_of(request)
-        target = str(body.get("target") or "")
-        if target == "_global":
-            if not global_layer.is_admin(settings, str(subject)):
-                raise HTTPException(status_code=403, detail="only an org admin may reset _global")
-            if not settings.global_system_workspace_path:
-                raise HTTPException(status_code=404, detail="no _global configured")
-            # write through the WORKSPACES-DIR mount (rw in dev) — the host-path mirror mount is ro
-            candidates = [Path(settings.workspaces_dir) / "_global", Path(settings.global_system_workspace_path)]
-            path = next((c for c in candidates if c.is_dir() and os.access(c, os.W_OK)), candidates[0])
-        elif target == "personal":
-            path = Path(wsr.workspace_dir(subject))
-        else:
-            raise HTTPException(status_code=400, detail="reset targets: personal | _global (shared workspaces are deleted; _system is never touched)")
-        if not path.is_dir():
-            raise HTTPException(status_code=404, detail="workspace not found")
-        seed = resolve_seed_dir(seeds_root=os.environ.get("VEXA_WORKSPACE_SEEDS_DIR"))
-        for child in path.iterdir():
-            if child.name == ".git":
-                continue
-            _sh.rmtree(child, ignore_errors=True) if child.is_dir() else child.unlink(missing_ok=True)
-        _sh.copytree(seed, path, dirs_exist_ok=True)
-        if (path / ".git").is_dir():
-            _sp.run(["git", "-C", str(path), "add", "-A"], check=False, capture_output=True)
-            _sp.run(["git", "-C", str(path), "-c", "user.name=vexa-platform", "-c", "user.email=platform@vexa.local",
-                     "commit", "-m", f"reseed {target}"], check=False, capture_output=True)
-        return {"target": target, "reset": True}
 
     # ── the COMPANY LAYER gate (PRD §9 decision 17; founder 2026-09-02) ──────────────────────────
     # A fresh instance serves nobody until an admin has written the thin company layer into
@@ -3425,395 +906,64 @@ def create_app(
             raise HTTPException(status_code=404, detail="the organisation tier is not present here")
         return target
 
-    @app.get("/api/global/state")
-    def global_state(request: Request):
-        """WHAT THE COMPANY LAYER HOLDS — the wizard's poll, and the honest answer to "why is this
-        instance still refusing people".
 
-        Readable by any authenticated subject on purpose: a non-admin who has just been refused at
-        the door deserves to be told the instance is mid-setup rather than that they are broken.
-        The company NAME is only returned once the gate is down — before that it is a half-written
-        answer to a question about somebody's employer."""
-        subject = subject_of(request)
-        gate = global_layer.instance_state(settings)
-        try:
-            st = global_layer.state(_global_store())
-        except HTTPException:
-            raise
-        except Exception as e:  # noqa: BLE001
-            raise HTTPException(status_code=500, detail=f"could not read the organisation tier: {e}")
-        down = gate.get("global_setup") == global_layer.COMPLETED
-        return {
-            "global_setup": gate.get("global_setup", global_layer.MISSING),
-            "company": (gate.get("company") or st["company"]) if down else None,
-            "present": st["present"],
-            "missing_files": st["missing_files"],
-            "reasons": st["reasons"],
-            "is_repo": st["is_repo"],
-            "commits": st["commits"],
-            "ready_to_accept": st["ready"],
-            "you_are_admin": global_layer.is_admin(settings, str(subject)),
-            "gate_sentence": global_layer.GATE_SENTENCE,
-        }
 
-    @app.post("/api/global/ready")
-    def global_ready(request: Request, body: dict = Body(default={})):
-        """ACCEPT the company layer: verify the files, commit them as the admin, lift the gate.
 
-        NOTHING MAY MARK ITSELF READY. The agent that wrote the layer asks for this verb and the
-        verb goes and looks — the five files present and non-empty, and a README that opens with
-        the company's name and one sentence of what it does. That last rule is the founder's:
-        *"the first chat needs to present itself knowing about itself — which company it's from and
-        what's their service."* An agent can only say which company it belongs to if a human wrote
-        the name down, so the gate does not lift on a README that does not carry one.
 
-        Admin-only, idempotent, and it reports WHY it refused rather than just refusing — the caller
-        is an agent mid-conversation with the one person who can fix it."""
-        subject = subject_of(request)
-        if not global_layer.is_admin(settings, str(subject)):
-            raise HTTPException(status_code=403,
-                                detail="only the instance admin may accept the company layer")
-        root = _global_store()
-        st = global_layer.state(root)
-        if not st["ready"]:
-            return JSONResponse(status_code=409, content={
-                "accepted": False,
-                "global_setup": global_layer.MISSING,
-                "missing_files": st["missing_files"],
-                "reasons": st["reasons"],
-                "next": "write the missing files into /workspaces/_global, then call this again",
-            })
-        email = str(body.get("author_email") or "").strip() or f"admin-{subject}@vexa.local"
-        name = str(body.get("author_name") or "").strip() or f"vexa admin {subject}"
-        try:
-            sha = global_layer.commit(root, author_email=email, author_name=name,
-                                      message=f"company layer: {st['company']}")
-        except Exception as e:  # noqa: BLE001
-            raise HTTPException(status_code=500, detail=f"could not commit the company layer: {e}")
-        try:
-            global_layer.mark_ready(settings, company=st["company"])
-        except Exception as e:  # noqa: BLE001
-            # The commit stands and the files are on disk; only the MARKER failed. Say exactly that
-            # — an agent told "failed" would rewrite files that are already correct.
-            raise HTTPException(status_code=502, detail=(
-                f"the company layer is committed ({sha}) but the instance gate could not be "
-                f"recorded: {e}"))
-        return {"accepted": True, "global_setup": global_layer.COMPLETED,
-                "company": st["company"], "service": st["service"], "commit": sha,
-                "files": st["present"]}
 
-    @app.post("/api/workspace/{workspace_id}/unshare")
-    def ws_unshare(workspace_id: str, request: Request):
-        """UN-SHARE a workspace (owner only) — move it back into the caller's PRIVATE store and drop every
-        member's index entry, so it stops being shared (mirror of share-enable). Returns the new private slug."""
-        subject = subject_of(request)
-        try:
-            membership_mod.require_role(wsr.root, workspace_id, subject, "owner")
-            members = membership_mod.read_members(wsr.root, workspace_id)
-            new_slug = ensure_workspace_private(wsr.root, subject, workspace_id)
-        except MembershipError as exc:
-            raise _member_error(exc)
-        except KeyError:
-            raise HTTPException(status_code=404, detail="workspace not found")
-        for m in members:  # best-effort: the shared workspace is gone, so drop the derived index entries
-            try:
-                mindex.remove(m.get("subject"), workspace_id)
-            except Exception:  # noqa: BLE001
-                pass
-        # The tree moved into the caller's private store and stopped being a group. Its id did NOT
-        # change — un-sharing is an administrative act, not a new workspace — so every link into it
-        # keeps resolving, and for everyone else it now answers `not-yours`, which is the truth.
-        _ws_sync(new_slug, kind="desk", owner=subject,
-                 ws_dir=workspace_slot_dir(wsr.root, subject, new_slug))
-        return {"slug": new_slug}
 
-    @app.post("/api/workspace/{slug}/share-enable")
-    def ws_share_enable(slug: str, request: Request):
-        """Make one of the caller's OWN workspaces shareable (promote a private workspace to a top-level
-        shared one if needed) and ensure the caller is its owner. Returns the shareable workspace_id — the
-        caller then mints invites against it. This is what lets ANY workspace be shared AFTER creation, with
-        no share-vs-not decision at create time."""
-        subject = subject_of(request)
-        try:
-            workspace_id, promoted = ensure_workspace_shareable(wsr.root, subject, slug)
-            if promoted:
-                membership_mod.ensure_owner(wsr.root, workspace_id, subject, index=mindex,
-                                            email=request.headers.get("x-user-email"), commit_fn=_pc)
-        except ValueError as exc:
-            raise HTTPException(status_code=400, detail=str(exc))
-        except KeyError:
-            raise HTTPException(status_code=404, detail="workspace not found")
-        except MembershipError as exc:
-            raise _member_error(exc)
-        return {"workspace_id": workspace_id, "promoted": promoted}
 
-    @app.post("/api/workspace/shared/new", status_code=201)
-    def ws_shared_new(request: Request, body: SharedNewBody = Body(default=SharedNewBody())):
-        """CREATE a new shared workspace and make the caller its OWNER — the bootstrap for the share flow.
-        A fresh top-level workspace (git-inited + seeded) is created at <root>/<workspace_id>; the caller is
-        granted owner in BOTH stores (policy/members.json + the index). The caller can then mint invites."""
-        subject = subject_of(request)
-        try:
-            wid = create_shared_workspace_dir(wsr.root, body.name)
-            membership_mod.ensure_owner(wsr.root, wid, subject, index=mindex,
-                                        email=request.headers.get("x-user-email"), commit_fn=_pc)
-            # The id is minted HERE, at creation, for the same reason it exists at all: this is the
-            # only moment the workspace's human NAME is known. `create_shared_workspace_dir`
-            # slugifies it into a directory and drops it, so without this line "ASWF DNA Project"
-            # never existed anywhere and the group could only ever be called
-            # `aswf-dna-project-b7b2ee`.
-            _ws_sync(wid, kind="group", name=(body.name or "").strip() or None, owner=subject)
-        except MembershipError as exc:
-            raise _member_error(exc)
-        except Exception as exc:  # noqa: BLE001 — surface a clean 500 (dir/seed failure) rather than a stack
-            logger.exception("shared-workspace create failed for subject=%s", subject)
-            raise HTTPException(status_code=500, detail="could not create shared workspace")
-        return {"workspace_id": wid, "role": "owner", "name": body.name}
 
-    @app.post("/api/workspace/invites", status_code=201)
-    def ws_invite_create(request: Request, body: InviteCreateBody = Body(...)):
-        """Mint a scoped invite token for a shared workspace. Auth: owner OR contributor of the target.
-        The workspace must be shareable (reserved/own-private refused). The token is returned ONCE; only
-        its hash is persisted in policy/invites.json."""
-        subject = subject_of(request)
-        try:
-            membership_mod.require_role(wsr.root, body.workspace_id, subject, "contributor")
-            minted = membership_mod.mint_invite(
-                wsr.root, body.workspace_id, role=body.role, created_by=subject,
-                expires_in_sec=body.expires_in_sec, max_uses=body.max_uses,
-                mode=body.mode, allowed_emails=body.allowed_emails, commit_fn=_pc,
-            )
-        except MembershipError as exc:
-            raise _member_error(exc)
-        # The client composes the accept URL; we hand back the token + id + terms once.
-        return {
-            "id": minted.id, "token": minted.token, "role": minted.role,
-            "workspace_id": body.workspace_id, "expires_at": minted.expires_at,
-            "max_uses": minted.max_uses, "mode": body.mode,
-            "accept_path": "/api/workspace/invites/accept",
-        }
 
-    @app.get("/api/workspace/invites/preview")
-    def ws_invite_preview(request: Request, token: str):
-        """READ-ONLY preview of an invite — the target workspace + terms — WITHOUT granting anything.
-        Powers the pre-join CONSENT screen: the invitee sees what the workspace is (its purpose), the role
-        they'd get, and who shared it BEFORE they log in / join. Capability-gated by the token (whoever
-        holds the link may preview it); no membership is checked or created, no use is consumed. 404 for a
-        token that matches nothing (never enumerates workspaces)."""
-        info = membership_mod.preview_invite(wsr.root, token)
-        if info is None:
-            raise HTTPException(status_code=404, detail="invalid invite")
-        wsid = info["workspace_id"]
-        # Human context for the card: the workspace's purpose + who shared it (their email when we've
-        # stored it — see the members roster; else the opaque subject as a last resort).
-        purpose = read_purpose(membership_mod._ws_dir(wsr.root, wsid))
-        shared_by = info.get("created_by")
-        for m in membership_mod.read_members(wsr.root, wsid):
-            if m.get("subject") == info.get("created_by") and m.get("email"):
-                shared_by = m["email"]
-                break
-        return {
-            "workspace_id": wsid, "name": wsid, "purpose": purpose,
-            "role": info["role"], "mode": info["mode"], "expires_at": info["expires_at"],
-            "shared_by": shared_by, "valid": info["valid"], "reason": info["reason"],
-        }
 
-    @app.post("/api/workspace/invites/accept")
-    def ws_invite_accept(request: Request, body: InviteAcceptBody = Body(...)):
-        """Redeem an invite token (any logged-in user) → membership in BOTH stores, use-count bumped.
-        Idempotent per user (accepting twice = one membership, no extra use consumed). The token carries
-        NO workspace id — we resolve it by scanning the shareable workspaces' invites for its hash.
-        Post-auth redeem (AMENDMENT 5): the caller is an already-authenticated user (X-User-Id); a
-        RESTRICTED invite additionally requires their VERIFIED email (X-User-Email, gateway-injected)
-        to be in the invite's allowed_emails."""
-        subject = subject_of(request)
-        # SECURITY BOUNDARY: X-User-Email is trusted as the caller's VERIFIED email ONLY because the
-        # gateway strips any client-sent x-user-email and re-injects the value it resolved from the
-        # api-key. That invariant holds solely when the gateway is agent-api's SOLE ingress. Today the
-        # terminal / host-local clients reach agent-api directly (no gateway hop), so on the direct edge
-        # this header is spoofable — restricted-mode invites are NOT a security boundary until agent-api
-        # is gateway-fronted (Stage 4). VEXA_REQUIRE_GATEWAY_IDENTITY (checked in subject_of) lets a
-        # hardened deploy reject non-gateway callers. See the TOPOLOGY BOUNDARY note in create_app.
-        subject_email = request.headers.get("x-user-email")
-        h = membership_mod.hash_token(body.token)
-        # Resolve which shared workspace this token belongs to by hash (never trust a client-declared id).
-        target_ws = None
-        root = wsr.root
-        for child in sorted(p for p in root.iterdir() if p.is_dir()) if root.exists() else []:
-            slug = child.name
-            if slug.startswith(".") or slug in membership_mod.RESERVED_SLUGS:
-                continue
-            for inv in membership_mod._read_json_list(child, membership_mod.INVITES_FILE):
-                if inv.get("hash") == h:
-                    target_ws = slug
-                    break
-            if target_ws:
-                break
-        if target_ws is None:
-            raise HTTPException(status_code=404, detail="invalid invite")
-        try:
-            result = membership_mod.accept_invite(
-                wsr.root, target_ws, token=body.token, subject=subject, subject_email=subject_email,
-                index=mindex, commit_fn=_pc,
-            )
-        except MembershipError as exc:
-            raise _member_error(exc)
-        return result
 
-    @app.delete("/api/workspace/invites/{invite_id}")
-    def ws_invite_revoke(invite_id: str, request: Request, workspace_id: str):
-        """Revoke an invite (owner/contributor of the workspace)."""
-        subject = subject_of(request)
-        try:
-            membership_mod.require_role(wsr.root, workspace_id, subject, "contributor")
-            membership_mod.revoke_invite(wsr.root, workspace_id, invite_id, commit_fn=_pc)
-        except MembershipError as exc:
-            raise _member_error(exc)
-        return {"ok": True, "invite_id": invite_id}
 
-    @app.get("/api/workspace/invites")
-    def ws_invites_list(request: Request, workspace_id: str):
-        """List a workspace's invites (owner/contributor). Hashes are never surfaced."""
-        subject = subject_of(request)
-        try:
-            membership_mod.require_role(wsr.root, workspace_id, subject, "contributor")
-            return {"invites": membership_mod.list_invites(wsr.root, workspace_id)}
-        except MembershipError as exc:
-            raise _member_error(exc)
 
-    @app.get("/api/workspace/members")
-    def ws_members_list(request: Request, workspace_id: str):
-        """List a workspace's members (owner/contributor). Opportunistically records the CALLER's own
-        verified email onto their member row (self-healing for members granted before emails were stored)
-        so the roster shows human labels, not opaque subject ids."""
-        subject = subject_of(request)
-        try:
-            membership_mod.require_role(wsr.root, workspace_id, subject, "contributor")
-            try:  # best-effort label refresh — never fail the list on a backfill hiccup
-                membership_mod.backfill_member_email(
-                    wsr.root, workspace_id, subject,
-                    request.headers.get("x-user-email"), commit_fn=_pc)
-            except Exception:  # noqa: BLE001
-                logger.debug("member email backfill skipped for %s in %s", subject, workspace_id, exc_info=True)
-            return {"members": membership_mod.read_members(wsr.root, workspace_id)}
-        except MembershipError as exc:
-            raise _member_error(exc)
 
-    @app.delete("/api/workspace/members/{member_subject}")
-    def ws_member_remove(member_subject: str, request: Request, workspace_id: str):
-        """Remove a member (owner only)."""
-        subject = subject_of(request)
-        try:
-            membership_mod.require_role(wsr.root, workspace_id, subject, "owner")
-            membership_mod.remove_member(wsr.root, workspace_id, member_subject, index=mindex, commit_fn=_pc)
-        except MembershipError as exc:
-            raise _member_error(exc)
-        return {"ok": True, "subject": member_subject}
-
-    @app.post("/api/workspace/members/{member_subject}/role")
-    def ws_member_role(member_subject: str, request: Request, workspace_id: str,
-                       body: RoleSetBody = Body(...)):
-        """Flip a member's role (owner only) — read <-> read/write permissions."""
-        subject = subject_of(request)
-        try:
-            membership_mod.require_role(wsr.root, workspace_id, subject, "owner")
-            rec = membership_mod.set_role(
-                wsr.root, workspace_id, member_subject, body.role,
-                changed_by=subject, index=mindex, commit_fn=_pc,
-            )
-        except MembershipError as exc:
-            raise _member_error(exc)
-        return rec
-
-    @app.post("/api/workspace/{workspace_id}/leave")
-    def ws_member_leave(workspace_id: str, request: Request):
-        """LEAVE a shared workspace — the caller removes THEMSELVES (any role; no owner gate). The
-        last-owner guard still applies: a sole creator must unshare or hand off ownership rather than
-        orphan the workspace, so their leave is refused (409) with that message."""
-        subject = subject_of(request)
-        if membership_mod.is_member(wsr.root, workspace_id, subject) is None:
-            raise HTTPException(status_code=404, detail="not a member of this workspace")
-        try:
-            membership_mod.remove_member(wsr.root, workspace_id, subject, index=mindex, commit_fn=_pc)
-        except MembershipError as exc:
-            raise _member_error(exc)
-        return {"ok": True, "left": workspace_id}
-
-    @app.get("/api/workspace/shared")
-    def ws_shared_list(request: Request):
-        """The "workspaces shared with me" listing, reconciled across BOTH membership stores.
-
-        ``users.data.memberships[]`` (the index) is the fast, cross-host read; ``policy/members.json``
-        is the authoritative one (Q6). Reading ONLY the mirror made every grant on this host invisible
-        whenever the internal edge to admin-api was unreachable — the route answered 200 with an empty
-        list, so a 403 on that hop rendered in the UI as "you have no shared workspaces" rather than as
-        an error. It is now a UNION, never a subtraction: an index row with no local dir is a workspace
-        that lives on another host and must still be listed, so the git store only ever ADDS rows the
-        index is missing. ``index_degraded`` says out loud when the mirror could not be read."""
-        subject = subject_of(request)
-        degraded = False
-        try:
-            rows = list(mindex.list(subject) or [])
-        except Exception as exc:  # noqa: BLE001 — the authoritative store still answers; never 500 this
-            logger.warning("membership index list failed for subject=%s: %s — serving policy/members.json",
-                           subject, exc)
-            rows, degraded = [], True
-        seen = {r.get("workspace_id") for r in rows}
-        for row in membership_mod.list_memberships(wsr.root, subject):
-            if row["workspace_id"] not in seen:
-                rows.append(row)
-        return {"memberships": rows, "index_degraded": degraded}
 
     # ── Settings → Models "Test" buttons (on-demand credential tests, fail-loud surface) ────────
     # Both test the caller's EFFECTIVE config — the same user > global > env resolution the
     # dispatch overlay / bot_spawn apply — so what's tested is what a turn/bot actually gets.
 
-    @app.get("/api/models/test")
-    def models_test(request: Request):
-        """Test the effective model credentials NOW: custom mode = a real 1-token completion
-        against the endpoint; subscription = mounted-credentials expiry check (the recurring
-        stale-Keychain 401 surfaces here with its remedy instead of at the next chat turn)."""
-        from control_plane import config_test as _ct
-        subject = subject_of(request)
-        cfg: dict = {}
-        mc = getattr(dispatcher, "_model_config", None)
-        if mc is not None:
-            try:
-                cfg = mc.resolve(subject) or {}
-            except Exception as exc:  # resolver down → still test the env floor, but SAY so
-                out = _ct.run_models_test({})
-                out["summary"] += f" (settings resolver unavailable: {exc} — tested env defaults)"
-                return out
-        return _ct.run_models_test(cfg)
 
-    @app.get("/api/transcription/test")
-    def transcription_test(request: Request):
-        """Probe the effective STT backend with its token (GET /balance): catches dead URLs,
-        rejected tokens, and the zero-balance-external-account case that 402s every segment."""
-        from control_plane import config_test as _ct
-        subject = subject_of(request)
-        url, token, source = "", "", "env"
-        settings = dispatcher.settings
-        admin = (settings.admin_api_url or "").rstrip("/")
-        if admin:  # same internal edge bot_spawn uses (bot-context carries the resolved override)
-            import urllib.request as _ur
-            try:
-                req = _ur.Request(f"{admin}/internal/users/{subject}/bot-context",
-                                  headers={"X-Internal-Secret":
-                                           settings.internal_api_secret.get_secret_value()})
-                with _ur.urlopen(req, timeout=5) as r:
-                    body = json.loads(r.read())
-                t = body.get("transcription") or {}
-                if t.get("url") or t.get("token"):
-                    url, token, source = t.get("url") or "", t.get("token") or "", "settings"
-            except Exception:
-                pass  # fall through to env — the probe result still says what was tested
-        if not url:
-            url = os.environ.get("TRANSCRIPTION_SERVICE_URL", "")
-            token = token or os.environ.get("TRANSCRIPTION_SERVICE_TOKEN", "")
-        elif not token:
-            token = os.environ.get("TRANSCRIPTION_SERVICE_TOKEN", "")
-        return _ct.run_transcription_test(url, token, source)
+
+    # ── THE ROUTES, BY OWNER ─────────────────────────────────────────────────────────────────
+    #
+    # `create_app` was 2,868 lines and 78 routes, and every lane that touched agent-api touched
+    # this one file — which is also where the identity hole lived (seam backlog B3). The routes
+    # now live in `control_plane/routers/`, one module per owner, and this function is what it
+    # says it is: build the app, build what the routes are built out of, include them.
+    #
+    # WHAT `build(**_deps)` IS FOR. Each handler closed over `create_app`'s locals — `wsr`,
+    # `settings`, `subject_of`, `_read_target` and twenty more. Passing them in a bag and rebinding
+    # each to THE NAME IT ALREADY HAD is what let every body move byte for byte: not one
+    # identifier inside a handler changed, so `git diff -M` reads as a move and a reviewer is
+    # reading the same code they reviewed before. Each router declares in its own `build()` which
+    # of these it takes, so "what does this router depend on" is answerable by reading one line.
+    #
+    # ORDER IS NOT LOAD-BEARING HERE, and that is checked rather than assumed: no two of these 78
+    # routes can match the same concrete URL under the same method (FastAPI resolves
+    # first-match-wins, so a pair that could would make the include order a behaviour). The check
+    # is `tests/test_route_table.py::test_no_two_routes_can_match_the_same_url`.
+    _deps = dict(
+        _clone_fn=_clone_fn, _compose_and_mint=_compose_and_mint,
+        _credential_refusal=_credential_refusal, _email_subject_lookup=_email_subject_lookup,
+        _entity_mounts=_entity_mounts, _friction_subject=_friction_subject,
+        _global_root=_global_root, _global_store=_global_store,
+        _internal_caller=_internal_caller, _manage_dir=_manage_dir,
+        _meeting_owner_lookup=_meeting_owner_lookup, _member_error=_member_error, _pc=_pc,
+        _read_target=_read_target, _repo=_repo, _require_shared_write=_require_shared_write,
+        _resolve_room=_resolve_room, _scaffold_is_for=_scaffold_is_for,
+        _scaffold_recipient_is=_scaffold_recipient_is, _scaffold_view=_scaffold_view,
+        _schedule_source=_schedule_source, _workspace_key=_workspace_key, _ws_here=_ws_here,
+        _ws_is_member=_ws_is_member, _ws_sync=_ws_sync, dispatcher=dispatcher,
+        friction=friction, invocations_url=invocations_url, live=live, mindex=mindex,
+        redis_url=redis_url, scaffolds=scaffolds, scheduler=scheduler, sess=sess,
+        settings=settings, stream_reader=stream_reader, subject_of=subject_of,
+        workspace_registry=workspace_registry, workspace_touches=workspace_touches, wsr=wsr)
+    for _r in (routers_health, routers_chats, routers_admin, routers_meetings, routers_scaffolds, routers_friction, routers_workspaces):
+        app.include_router(_r.build(**_deps))
+
     return app
